% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{aux_config}
\alias{aux_config}
\alias{convert_to_matrix}
\alias{aux_config}
\alias{matrix_to_long}
\alias{aux_config}
\alias{join_data_matrices}
\title{Auxiliary functions for data matrix manipulation}
\usage{
convert_to_matrix(df_long, feature_id_column = "peptide_group_label",
  measure_column = "Intensity", sample_id_col = "FullRunName")

matrix_to_long(data_matrix, sample_annotation,
  feature_id_column = "peptide_group_label", measure_col = "Intensity",
  step = NA)

join_data_matrices(matrix_list, step, sample_annotation,
  measure_column = "Intensity")
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single sample,
thus it has minimally, `sample_id_col`, `feature_id_column` and `measure_column`, but usually also `m_score` (in OpenSWATH output result file)}

\item{feature_id_column}{name of the column with feature/gene/peptide/protein ID used with long format matrix (`df_long`). In wide format (`data_matrix`) this would be the row name}

\item{measure_column}{if `df_long` is among the parameters, it is the column with expression/abundance/intensity,
otherwise, it is used internally for consistency}

\item{sample_id_col}{name of the column in sample_annotation file,
where the filenames (colnames of the data matrix are found)}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix,
with feature IDs in rownames and file/sample names as colnames.
Usually the log transformed version of the original data}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be repeated as row names) 2) biological and 3) technical covariates (batches etc)}

\item{step}{normalization step (e.g. `Raw` or `Quantile_normalized` or `qNorm_ComBat`).
Useful if consecutive steps are compared in plots.
Note that in plots these are usually ordered alphabetically, so it's worth naming with numbers, e.g. `1_raw`, `2_quantile`}

\item{matrix_list}{list of matrices in `data_matrix` format}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be repeated as row names) 2) biological and 3) technical covariates (batches etc)}
}
\value{
`data_matrix`-like matrix (features in rows, samples in columns)

`df_long`-like object

`df_long` like data frame with a row, having entries for
1) `feature_id_col` (e.g. peptide name), 2) `sample_id_col` (e.g. filename),
3) `measure_col` (e.g. intensity/expression) and 4) `step` (e.g. `raw`, `quantile_norm`)
}
\description{
Converting from long to wide (matrix), wide (matrix) to long, joining the data matrices
}
