% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{calculate_sample_corr_distribution}
\alias{calculate_sample_corr_distribution}
\title{Calculates correlation of data matrix and calculates correlation distribution for all pairs 
of the replicated samples}
\usage{
calculate_sample_corr_distribution(data_matrix, repeated_samples,
  sample_annotation, biospecimen_id_col, sample_id_col, batch_col)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{repeated_samples}{if \code{NULL}, only repeated sample correlation is plotted}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{biospecimen_id_col}{column in \code{sample_annotation} 
that defines a unique bio ID, which is usually a 
combination of conditions or groups.
 Tip: if such ID is absent, but can be defined from several columns,
 create new \code{biospecimen_id} column}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}
}
\value{
dataframe with the following columns, that 
are suggested to use for plotting in 
\code{\link{plot_sample_corr_distribution}} as \code{plot_param}:
\enumerate{
\item \code{replicate}
\item \code{batch_the_same}
\item \code{batch_replicate}
\item \code{batches}
}
other columns are: \enumerate{
\item \code{sample_id_1} & \code{sample_id_2}, both 
generated from \code{sample_id_col} variable
\item \code{correlation} - correlation of two corresponding samples
\item \code{batch_1} & \code{batch_2} or analogous, 
created the same as \code{sample_id_1}
}
}
\description{
Calculates correlation of data matrix and calculates correlation distribution for all pairs 
of the replicated samples
}
