% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effects_old.R
\name{center_feature_batch_medians_df}
\alias{center_feature_batch_medians_df}
\title{DEPRECATED: center_feature_batch_medians_df}
\usage{
center_feature_batch_medians_df(
  df_long,
  sample_annotation = NULL,
  sample_id_col = "FullRunName",
  batch_col = "MS_batch",
  feature_id_col = "peptide_group_label",
  measure_col = "Intensity",
  keep_all = "default",
  no_fit_imputed = TRUE,
  qual_col = NULL,
  qual_value = NULL
)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col}
and a \code{measure_col}, but usually also an \code{m_score} (in OpenSWATH
output result file). See \code{help("example_proteome")} for more details.}

\item{sample_annotation}{data frame with:
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names)
  \item biological covariates
  \item technical covariates (batches etc) }.
  See \code{help("example_sample_annotation")}}

\item{sample_id_col}{name of the column in \code{sample_annotation} table,
where the filenames (colnames of the \code{data_matrix} are found).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison (or other, non-batch factor to be mapped to color in plots).}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{keep_all}{when transforming the data (normalize, correct) - acceptable
values: all/default/minimal (which set of columns be kept).}

\item{no_fit_imputed}{(logical) whether to use imputed (requant) values, as flagged in
\code{qual_col} by \code{qual_value} for data transformation}

\item{qual_col}{column to color point by certain value denoted
by \code{qual_value}. Design with inferred/requant values in
OpenSWATH output data,
which means argument value has to be set to \code{m_score}.}

\item{qual_value}{value in \code{qual_col} to color. For OpenSWATH data,
this argument value has to be set to \code{2} (this is an \code{m_score}
value for imputed values (requant values).}
}
\value{
A data.frame in long format with batch-corrected values in \code{measure_col}
  and original values preserved in \code{preBatchCorr_[measure_col]}.
}
\description{
Use [center_feature_batch()] with `format="long", stat="medians"`.
}
