% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_proteome.R
\name{clean_proteome}
\alias{clean_proteome}
\alias{clean_requants}
\alias{clean_proteome}
\alias{remove_peptides_with_missing_batch}
\alias{clean_proteome}
\alias{summarize_peptides}
\title{Remove peptides with too many missing values}
\usage{
clean_requants(df_long, sample_annotation, batch_column = "MS_batch.final",
  feature_id_column = "peptide_group_label", threshold_batch = 0.3,
  threshold_global = 0.3)

remove_peptides_with_missing_batch(df_long, batch_column = "MS_batch.final",
  feature_id_column = "peptide_group_label")

summarize_peptides(df_long, sample_id_col = "FullRunName",
  feature_id_column = "peptide_group_label")
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single sample,
thus it has minimally, `sample_id_col`, `feature_id_column` and `measure_column`, but usually also `m_score` (in OpenSWATH output result file)}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be repeated as row names) 2) biological and 3) technical covariates (batches etc)}

\item{batch_column}{column in `sample_annotation` that should be used for batch comparison}

\item{feature_id_column}{name of the column with feature/gene/peptide/protein ID used with long format matrix (`df_long`). In wide format (`data_matrix`) this would be the row name}

\item{threshold_batch}{maximal fraction of missing values for a feature for a batch}

\item{threshold_global}{maximal fraction of missing values for a feature globally}

\item{sample_id_col}{name of the column in sample_annotation file,
where the filenames (colnames of the data matrix are found)}

\item{measure_column}{if `df_long` is among the parameters, it is the column with expression/abundance/intensity,
otherwise, it is used internally for consistency}
}
\value{
`df_long`-like data frame with the requant values removed completely and peptides,
missing in batch or globally beyond threshold are removed

`df_long`-like data frame free of peptides that were not detected in each batch
}
\description{
Remove peptides with too many missing values

Remove peptides that are missing in the whole batch

summarize peptides by sample (ranking) and on the contrary, across peptide-wise across samples
}
\details{
useful for some downstream functions as ComBat normalization, that would not work otherwise
}
