% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{cluster_samples}
\alias{cluster_samples}
\title{cluster the data matrix to visually inspect which confounder dominates}
\usage{
cluster_samples(data_matrix, color_df, plot_title, distance = "euclidean",
  agglomeration = "complete", ...)
}
\arguments{
\item{title}{}
}
\description{
cluster the data matrix to visually inspect which confounder dominates
}
