% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{color_list_to_df}
\alias{color_list_to_df}
\title{Color list to data frame}
\usage{
color_list_to_df(color_list, sample_annotation, sample_id_col = "FullRunName")
}
\arguments{
\item{color_list}{list of colors}

\item{sample_annotation}{factor-based configuration
of the sample annotation}

\item{sample_id_col}{Column containing sample identifiers. When absent or
unusable, row names and then row positions are used.}
}
\value{
a data frame representation of the input color list
}
\description{
Turn color list to df (to use in the hierarchical clustering)
}
\keyword{internal}
