% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{convert_annotation_classes}
\alias{convert_annotation_classes}
\title{Convert factor and numeric columns}
\usage{
convert_annotation_classes(df, factor_columns, numeric_columns)
}
\arguments{
\item{df}{data frame with sample annotations.}

\item{factor_columns}{character vector of factor columns.}

\item{numeric_columns}{character vector of numeric columns.}
}
\value{
data frame with converted columns.
}
\description{
Convert specified factor columns to factor type and
numeric columns to numeric.
}
\examples{
annotation <- data.frame(
    batch = c("A", "B"),
    injection_order = c("1", "2")
)
converted <- convert_annotation_classes(
    annotation,
    factor_columns = "batch",
    numeric_columns = "injection_order"
)
vapply(converted, class, character(1))
}
