% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{Convert from long data frame to data matrix (features in rows, samples in columns)}
\usage{
convert_to_matrix(data_df_long, feature_id_column = "peptide_group_label",
  measure_column = "Intensity", sample_id_column = "FullRunName")
}
\arguments{
\item{sample_id_column}{}
}
\description{
Convert from long data frame to data matrix (features in rows, samples in columns)
}
