% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effect.R
\name{correct_batch}
\alias{correct_batch}
\alias{center_peptide_batch_medians}
\alias{adjust_batch_trend}
\alias{correct_with_ComBat}
\title{Correction of batch effects in the data}
\usage{
center_peptide_batch_medians(df_long, sample_annotation = NULL,
  sample_id_col = "FullRunName", batch_col = "MS_batch",
  feature_id_col = "peptide_group_label", measure_col = "Intensity")

adjust_batch_trend(data_matrix, sample_annotation,
  batch_col = "MS_batch", feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName", measure_col = "Intensity",
  sample_order_col = "order", fit_func = fit_nonlinear,
  abs.threshold = 5, pct.threshold = 0.2, ...)

correct_with_ComBat(data_matrix, sample_annotation,
  sample_id_col = "FullRunName", batch_col = "MS_batch",
  par.prior = TRUE)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{measure_col}{if `df_long` is among the parameters, it is the column
with expression/abundance/intensity, otherwise, it is used internally for
consistency}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{sample_order_col}{column, determining the order of sample MS run, used
as covariate to fit the non-linear fit}

\item{fit_func}{function to fit the (non)-linear trend}

\item{...}{other parameters, usually those of the `fit_func`}

\item{par.prior}{whether parametrical or non-parametrical prior should be used}

\item{return_long}{whether the result should be the "long" data frame (as
`df_long`) or "wide" (as `data_matrix`)}
}
\value{
`data_matrix`-size data matrix with batch-effect corrected with
  per-feature batch median centering

list of two items: 1) `data_matrix`, adjusted with continious fit; 
2) fit_df, used to examine the fitting curves

`data_matrix`-size data matrix with batch-effect corrected by
  `ComBat`
}
\description{
Correction of batch effects in the data

Median centering of the peptides (per batch median)

adjust batch signal trend with the custom (continuous) fit

Standardized input-output ComBat normalization ComBat allows users to adjust
for batch effects in datasets where the batch covariate is known, using
methodology described in Johnson et al. 2007. It uses either parametric or
non-parametric empirical Bayes frameworks for adjusting data for batch
effects.  Users are returned an expression matrix that has been corrected for
batch effects. The input data are assumed to be cleaned and normalized before
batch effect removal.
}
\examples{

}
\seealso{
\code{\link{fit_nonlinear}}
}
