% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effect.R
\name{correct_batch_trend}
\alias{correct_batch_trend}
\title{Batch correction method allows correction of continuous sigal drift within batch and 
discrete difference across batches.}
\usage{
correct_batch_trend(data_matrix, sample_annotation,
  fitFunc = "loess_regression", discreteFunc = "MedianCentering",
  batch_col = "MS_batch", feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName", measure_col = "Intensity",
  sample_order_col = "order", loess.span = 0.75, abs.threshold = 5,
  pct.threshold = 0.1, ...)
}
\arguments{
\item{discreteFunc}{function to use for discrete batch correction (`MedianCentering` or `ComBat`)}

\item{...}{other parameters, usually of `normalize_custom_fit`, and `fit_func`}

\item{fitFunct}{function to use for the fit (currently only `loess_regression` available)}
}
\value{
`data_matrix`-size data matrix with batch-effect corrected by fit and discrete functions
}
\description{
Batch correction method allows correction of continuous sigal drift within batch and 
discrete difference across batches.
}
