% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_ComBat.R
\name{correct_with_mComBat}
\alias{correct_with_mComBat}
\title{The modification to ComBat by Johson, Rabinovic, and Li (Biostatistics, 2007)
to shift data to mean and variance of subset of samples rather
than grand mean and pooled variance (http://www.biomedcentral.com/1471-2105/16/63).
The m-ComBat method was originally developed by Caleb K. Stein in 2014 (https://github.com/SteinCK/M-ComBat)
and have been adapted to use inside proBatch package.
Reference:
  Stein, C.K., Qu, P., Epstein, J. et al. Removing batch effects from
  purified plasma cell gene expression microarrays with modified ComBat.
  BMC Bioinformatics 16, 63 (2015). https://doi.org/10.1186/s12859-015-0478-3}
\usage{
correct_with_mComBat(
  x,
  sample_annotation = NULL,
  feature_id_col = "peptide_group_label",
  measure_col = "Intensity",
  sample_id_col = "FullRunName",
  batch_col = "MS_batch",
  format = c("long", "wide"),
  covariates_cols = NULL,
  fill_the_missing = NULL,
  keep_all = "default",
  no_fit_imputed = TRUE,
  qual_col = NULL,
  qual_value = NULL,
  mComBat_center = NULL
)
}
\arguments{
\item{x}{Data in long (\code{data.frame}) or wide (\code{matrix}) form, controlled by \code{format}.}

\item{sample_annotation}{data frame with:
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names)
  \item biological covariates
  \item technical covariates (batches etc) }.
  See \code{help("example_sample_annotation")}}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{sample_id_col}{name of the column in \code{sample_annotation} table,
where the filenames (colnames of the \code{data_matrix} are found).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison (or other, non-batch factor to be mapped to color in plots).}

\item{format}{One of \code{"long"} or \code{"wide"}.}

\item{covariates_cols}{optional sample_annotation columns for the design matrix.
These covariates will be preserved during batch correction.}

\item{fill_the_missing}{Missing-value policy prior to ComBat. If \code{NULL}, no action is taken
and the call will \emph{fail if the matrix contains NA}. Set \code{FALSE} to drop rows with NA,
or provide a numeric value to impute (use with caution).}

\item{keep_all}{For long format, columns to retain (see \code{subset_keep_cols()}).}

\item{no_fit_imputed}{If \code{TRUE} and \code{qual_col} provided, masked values are
excluded when building the matrix (original values still corrected).}

\item{qual_col}{column to color point by certain value denoted
by \code{qual_value}. Design with inferred/requant values in
OpenSWATH output data,
which means argument value has to be set to \code{m_score}.}

\item{qual_value}{value in \code{qual_col} to color. For OpenSWATH data,
this argument value has to be set to \code{2} (this is an \code{m_score}
value for imputed values (requant values).}

\item{mComBat_center}{If using \code{use_mComBat=TRUE}, the center which should be used
to center the data (https://doi.org/10.1186/s12859-015-0478-3).}
}
\value{
Matrix if \code{format="wide"}, data.frame if \code{"format="long"} with batch effects removed
}
\description{
Removes batch-associated linear effects for plotting/unsupervised
  tasks. Works for long or wide via \code{format}. Use \code{covariates_cols}
  to keep biological effects in the design (not removed).
}
\details{
' @title Batch effect correction with modified ComBat (m-ComBat)
}
\examples{
data(
    list = c("example_sample_annotation", "example_proteome_matrix"),
    package = "proBatch"
)
# Wide format
corrected_wide <- correct_with_mComBat(
    x = example_proteome_matrix,
    sample_annotation = example_sample_annotation,
    batch_col = "batch",
    covariates_cols = c("condition"),
    feature_id_col = "peptide_group_label",
    format = "wide",
    na_action = FALSE
)
head(corrected_wide)
}
\seealso{
\code{\link[limma]{removeBatchEffect}}
}
