% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{date_to_order}
\alias{date_to_order}
\title{Add order and dateTime column to sample annotation data frame}
\usage{
date_to_order(sample_annotation, time_column, dateTimeFormat = c("\%b_\%d",
  "\%H:\%M:\%S"), order_column = "order")
}
\arguments{
\item{order_column}{}
}
\description{
Add order and dateTime column to sample annotation data frame
}
