% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{define_batches_by_MS_pauses}
\alias{define_batches_by_MS_pauses}
\title{Identify stretches of time between runs that are long and split a batches by them}
\usage{
define_batches_by_MS_pauses(date_vector, threshold, minimal_batch_size = 5,
  batch_name = "MS_batch")
}
\arguments{
\item{batch_name}{}
}
\description{
Identify stretches of time between runs that are long and split a batches by them
}
