% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{define_batches_by_MS_pauses_within_instrument}
\alias{define_batches_by_MS_pauses_within_instrument}
\title{Batch by date/time}
\usage{
define_batches_by_MS_pauses_within_instrument(date_vector, threshold,
  minimal_batch_size = 5, batch_name = "MS_batch")
}
\arguments{
\item{date_vector}{POSIX or numeric-like vector corresponding to the sample
MS profile acquisition timepoint}

\item{threshold}{time difference that would mean there was an interruption}

\item{minimal_batch_size}{minimal number of samples in a batch}

\item{batch_name}{string with a self-explanatory name for the batch (e.g.
`MS_batch` for MS-proteomics) to which batch number will be added}
}
\value{
vector of batches for each sample
}
\description{
Identify long stretches of time between samples and split them into batches.
Most users are going to want to call define_batches_by_MS_pauses, rather then
this function
}
