% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{distribution_of_cor}
\alias{distribution_of_cor}
\title{Plot distribution of correlations}
\usage{
distribution_of_cor(data_matrix_sub, facet_var = NULL, theme = "classic")
}
\arguments{
\item{data_matrix_sub}{}
}
\description{
Plot distribution of correlations
}
