% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{.corr_distribution}
\alias{.corr_distribution}
\title{Calculates correlation of data matrix and calculates correlation distribution for all pairs 
of the replicated samples}
\usage{
.corr_distribution(data_matrix, repeated_samples, sample_annotation,
  biospecimen_id_col, sample_id_col, batch_col)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{repeated_samples}{if `NULL`, only repeated sample correlation is plotted}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be
repeated as row names) 2) biological and 3) technical covariates (batches
etc)}

\item{biospecimen_id_col}{column in `sample_annotation` 
that defines a unique bio ID, which is usually a 
combination of conditions or groups.
 Tip: if such ID is absent, but can be defined from several columns,
 create new \code{biospecimen_id} column}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix) are found}

\item{batch_col}{column in `sample_annotation` that should be used for
batch comparison}
}
\value{
dataframe with the following columns, that 
are suggested to use for plotting in 
\code{\link{plot_sample_corr_distribution}} as \code{plot_param}:
\enumerate{
\item \code{replicate}
\item \code{batch_the_same}
\item \code{batch_replicate}
\item \code{batches}
}
other columns are: \enumerate{
\item \code{sample_id_1} & \code{sample_id_2}, both 
generated from \code{sample_id_col} variable
\item \code{correlation} - correlation of two corresponding samples
\item \code{batch_1} & \code{batch_2} or analogous, 
created the same as \code{sample_id_1}
}
}
\description{
Calculates correlation of data matrix and calculates correlation distribution for all pairs 
of the replicated samples
}
\keyword{internal}
