% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{.corr_distribution_prot}
\alias{.corr_distribution_prot}
\title{Transform square correlation matrix into long data frame of correlations}
\usage{
.corr_distribution_prot(data_matrix, peptide_annotation, protein_col,
  feature_id_col)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{peptide_annotation}{df with petpides and 
their corresponding proteins}

\item{protein_col}{the column name in \code{peptide_annotation} 
with protein names}

\item{feature_id_col}{column in \code{peptide_annotation} that 
captures peptide names are found}
}
\value{
dataframe with peptide correlation coefficients 
that are suggested to use for plotting in 
\code{\link{plot_peptide_corr_distribution}} as \code{plot_param}:
}
\description{
Transform square correlation matrix into long data frame of correlations
}
\keyword{internal}
