% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_sample_annotation}
\alias{example_sample_annotation}
\title{Sample annotation data version 1}
\format{
A data frame with 233 rows and 11 variables:
\describe{
  \item{FullRunName}{name of the file with the measurement for each sample,
  referred to as \code{sample_id_col}}
  \item{MS_batch}{mass-spectrometry batch: 4-level factor of
  manually annotated batches}
  \item{EarTag}{mouse ID, i.e. ID of the biological object. Only 14 mice have
  been replicated, one mouse was profiled 7 times.}
  \item{Strain}{mouse strain ID from BXD population set -
  biological covariate #1, 51 Strain represented}
  \item{Diet}{diet, biological covariate #2 - either
  \code{HFD} = `High Fat Diet` or \code{CD} = `Chow Diet`}
  \item{Sex}{mice sex - biological covariate #3}
  \item{RunDate}{mass-spectrometry running date. In combination
  with \code{RunTime} used for running order determination. Vector of class
  \code{"difftime"} and \code{"hms"}}
  \item{RunTime}{mass-spectrometry running time. In combination
  with \code{RunDate} used for running order determination.Vector of class
  \code{"POSIXct"} and \code{"POSIXt"}}
  \item{DateTime}{numeric date and time generated by
  \code{date_to_sample_order}}
  \item{order}{order of samples generated by sorting DateTime
  in \code{date_to_sample_order}}
  \item{digestion_batch}{peptide digestion batch: 4-level
  factor of manually annotated batches} ...
}
}
\value{
A data frame with 233 rows and 11 variables.
}
\description{
This is data from BXD mouse population aging study with mock instruments to
show how instrument-specific functionality works
}
\examples{
data("example_sample_annotation", package = "proBatch")
head(example_sample_annotation)
}
