% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{feature_level_diagnostics}
\alias{feature_level_diagnostics}
\alias{plot_single_feature}
\alias{plot_peptides_of_one_protein}
\alias{plot_spike_in}
\alias{plot_iRT}
\alias{plot_with_fitting_curve}
\title{Plotting peptide measurements}
\usage{
plot_single_feature(
  feature_name,
  df_long,
  sample_annotation = NULL,
  pbf_name = NULL,
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  feature_id_col = "peptide_group_label",
  geom = c("point", "line"),
  qual_col = NULL,
  qual_value = NULL,
  batch_col = "MS_batch",
  color_by_batch = FALSE,
  color_scheme = "brewer",
  order_col = "order",
  vline_color = "red",
  facet_col = NULL,
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = NULL,
  theme = "classic",
  ylimits = NULL,
  base_size = 20
)

plot_peptides_of_one_protein(
  protein_name,
  peptide_annotation = NULL,
  protein_col = "ProteinName",
  df_long,
  sample_annotation = NULL,
  pbf_name = NULL,
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  feature_id_col = "peptide_group_label",
  geom = c("point", "line"),
  qual_col = NULL,
  qual_value = NULL,
  batch_col = "MS_batch",
  color_by_batch = FALSE,
  color_scheme = "brewer",
  order_col = "order",
  vline_color = "red",
  facet_col = NULL,
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = sprintf("Peptides of \%s protein", protein_name),
  theme = "classic",
  base_size = 20
)

plot_spike_in(
  spike_ins = "BOVIN",
  peptide_annotation = NULL,
  protein_col = "ProteinName",
  df_long,
  sample_annotation = NULL,
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  feature_id_col = "peptide_group_label",
  geom = c("point", "line"),
  qual_col = NULL,
  qual_value = NULL,
  batch_col = "MS_batch",
  color_by_batch = FALSE,
  color_scheme = "brewer",
  order_col = "order",
  vline_color = "red",
  facet_col = NULL,
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = sprintf("Spike-in \%s plots", spike_ins),
  theme = "classic",
  base_size = 20
)

plot_iRT(
  irt_pattern = "iRT",
  peptide_annotation = NULL,
  protein_col = "ProteinName",
  df_long,
  sample_annotation = NULL,
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  feature_id_col = "peptide_group_label",
  geom = c("point", "line"),
  qual_col = NULL,
  qual_value = NULL,
  batch_col = "MS_batch",
  color_by_batch = FALSE,
  color_scheme = "brewer",
  order_col = "order",
  vline_color = "red",
  facet_col = NULL,
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = "iRT peptide profile",
  theme = "classic",
  base_size = 20
)

plot_with_fitting_curve(
  feature_name,
  fit_df,
  fit_value_col = "fit",
  df_long,
  sample_annotation = NULL,
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  feature_id_col = "peptide_group_label",
  geom = c("point", "line"),
  qual_col = NULL,
  qual_value = NULL,
  batch_col = "MS_batch",
  color_by_batch = FALSE,
  color_scheme = "brewer",
  order_col = "order",
  vline_color = "grey",
  facet_col = NULL,
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = sprintf("Fitting curve of \%s peptide", paste(feature_name, collapse =
    " ")),
  theme = "classic",
  base_size = 20
)
}
\arguments{
\item{feature_name}{name of the selected feature (e.g. peptide) for
diagnostic profiling}

\item{df_long}{data frame where each row is a single feature in a single
sample, or a `ProBatchFeatures` object.}

\item{sample_annotation}{data frame with:
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names)
  \item biological covariates
  \item technical covariates (batches etc) }.
  See \code{help("example_sample_annotation")}}

\item{pbf_name}{Assay name used when `df_long` is a `ProBatchFeatures`
object (used by `plot_single_feature()` and
`plot_peptides_of_one_protein()`). If `NULL`, [pb_current_assay()] is used.}

\item{sample_id_col}{name of the column in \code{sample_annotation} table,
where the filenames (colnames of the \code{data_matrix} are found).}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this
corresponds to the row names.}

\item{geom}{whether to show the feature as points and/or connect by lines
(accepted values are: 1. \code{point}, \code{line} and
\code{c('point', 'line')})}

\item{qual_col}{column to color point by certain value denoted
by \code{qual_value}. Design with inferred/requant values in OpenSWATH
output data, which means argument value has to be set to \code{m_score}.}

\item{qual_value}{value in \code{qual_col} to color. For OpenSWATH data,
this argument value has to be set to \code{2} (this is an \code{m_score}
value for imputed values (requant values).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison (or other, non-batch factor to be mapped to color in plots).}

\item{color_by_batch}{(logical) whether to color points and connecting lines
by batch factor as defined by \code{batch_col}.}

\item{color_scheme}{a named vector of colors to map to \code{batch_col},
names corresponding to the levels of the factor. For continuous variables,
vector doesn't need to be named.}

\item{order_col}{column in \code{sample_annotation} that determines sample
order. It is used for in initial assessment plots
(\link{plot_sample_mean_or_boxplot}) and  feature-level diagnostics
(\link{feature_level_diagnostics}). Can be `NULL`
   if sample order is irrelevant (e.g. in genomic experiments). For more
   details, order definition/inference, see \link{define_sample_order} and
   \link{date_to_sample_order}}

\item{vline_color}{color of vertical lines, typically separating
different MS batches in ordered runs; should be `NULL` for
experiments without intrinsic order}

\item{facet_col}{column  in \code{sample_annotation} with a batch factor to
separate plots into facets; usually 2nd to \code{batch_col}. Most meaningful
for multi-instrument MS experiments (where each instrument has its own
order-associated effects (see \code{order_col}) or simultaneous examination
of two batch factors (e.g. preparation day and measurement day). For
single-instrument case should be set to `NULL`}

\item{filename}{path where the results are saved.
If null the object is returned to the active window; otherwise, the object
is save into the file. Currently only pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image height}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{theme}{ggplot theme, by default \code{classic}.
Can be easily overridden}

\item{ylimits}{range of y-axis to plot feature-level trends}

\item{base_size}{base font size}

\item{protein_name}{name of the protein as defined in \code{protein_col}}

\item{peptide_annotation}{long format data frame with peptide ID and their
corresponding protein and/or gene annotations. See
\code{help("example_peptide_annotation")}.}

\item{protein_col}{column where protein names are specified}

\item{spike_ins}{name of feature(s), typically proteins that were spiked in
for control}

\item{irt_pattern}{substring used to identify iRT proteins in
\code{protein_col}}

\item{fit_df}{data frame output of \code{adjust_batch_trend_df} to be plotted
with the line}

\item{fit_value_col}{column in \code{fit_df} where the values for fitting
trend are found}
}
\value{
ggplot2 type plot of \code{measure_col} vs \code{order_col},
  faceted by \code{feature_name} and (optionally) by \code{facet_col}
}
\description{
Creates a peptide faceted ggplot2 plot of the value in
\code{measure_col} vs \code{order_col} (if `NULL`, x-axis is simply a sample
name order). Additionally, the resulting plot can also be colored either by
batch factor, by quality factor (e.g. imputed/non-imputed) and, if needed,
faceted by another batch factor, e.g. an instrument.
 If the non-linear curve was fit, this can also be added to the plot, see
 functions specific to each case below
}
\examples{
data(list = c(
    "example_sample_annotation", "example_proteome",
    "example_peptide_annotation"
), package = "proBatch")

sample_annotation <- example_sample_annotation

feature_ids <- c(
    "10231_QDVDVWLWQQEGSSK_2", "10768_RLESELDGLR_2",
    "10062_NVGVSFYADKPEVTQEQK_3", "10063_NVGVSFYADKPEVTQEQKK_3",
    "1146_ADVTPADFSEWSK_3", "12476_TPVISGGPYEYR_2"
)
peptide_annotation <- subset(
    example_peptide_annotation, peptide_group_label \%in\% feature_ids)
proteome <- subset(
    example_proteome, peptide_group_label \%in\% feature_ids)
feature_id <- feature_ids[1]

feature_plot <- plot_single_feature(
    feature_name = feature_id,
    df_long = proteome,
    sample_annotation = sample_annotation,
    color_by_batch = TRUE,
    batch_col = "MS_batch"
)

protein_plot <- plot_peptides_of_one_protein(
    protein_name = "Haao",
    peptide_annotation = peptide_annotation,
    df_long = proteome,
    sample_annotation = sample_annotation,
    protein_col = "Gene"
)

spike_in_plot <- plot_spike_in(
    spike_ins = "BOVINE_A1ag",
    peptide_annotation = peptide_annotation,
    df_long = proteome,
    sample_annotation = sample_annotation,
    protein_col = "Gene"
)

irt_plot <- plot_iRT(
    irt_pattern = "iRT",
    peptide_annotation = peptide_annotation,
    df_long = proteome,
    sample_annotation = sample_annotation,
    protein_col = "Gene"
)

fit_input <- adjust_batch_trend_df(
    proteome[proteome$peptide_group_label == feature_id, ],
    sample_annotation,
    span = 0.7
)

curve_plot <- plot_with_fitting_curve(
    feature_name = feature_id,
    df_long = proteome,
    sample_annotation = sample_annotation,
    fit_df = fit_input
)

}
