% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{generate_colors_for_numeric}
\alias{generate_colors_for_numeric}
\title{Generates color vector from continuous palette}
\usage{
generate_colors_for_numeric(palette_type = "brewer", i = 1)
}
\arguments{
\item{palette_type}{'brewer' or 'viridis'}

\item{i}{if \code{palette_type} is 'brewer' the palette argument to
\code{brewer_pal}. If \code{palette_type} is 'viridis' the option argument
to \code{viridis_pal}}
}
\value{
vector of colors
}
\description{
Generates a vector of colors for a vector of numeric, POSIXct (i.e. the
(signed) number of seconds since the beginning of 1970 , or factors
}
\examples{
generate_colors_for_numeric("brewer", i = 1)

}
\keyword{internal}
