% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{genome_wide_diagnostics}
\alias{genome_wide_diagnostics}
\alias{plot_sample_mean}
\title{Set of tools for genome-wide diagnostics of batch effects}
\usage{
plot_sample_mean(data_matrix, sample_annotation,
  sample_id_col = "FullRunName", order_column = "order",
  batch_column = NULL, color_by_batch = F, theme = "classic",
  title = NULL, color_scheme = "brewer")
}
\arguments{
\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be repeated as row names) 2) biological and 3) technical covariates (batches etc)}

\item{sample_id_col}{name of the column in sample_annotation file,
where the filenames (colnames of the data matrix are found)}

\item{batch_column}{column in `sample_annotation` that should be used for batch comparison}

\item{color_scheme}{}

\item{measure_column}{if `df_long` is among the parameters, it is the column with expression/abundance/intensity,
otherwise, it is used internally for consistency}

\item{df_long}{data frame where each row is a single feature in a single sample,
thus it has minimally, `sample_id_col`, `feature_id_column` and `measure_column`, but usually also `m_score` (in OpenSWATH output result file)}

\item{feature_id_column}{name of the column with feature/gene/peptide/protein ID used with long format matrix (`df_long`). In wide format (`data_matrix`) this would be the row name}
}
\value{
ggplot2 class plot of average intensities
}
\description{
Set of tools for genome-wide diagnostics of batch effects
}
