% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{get_peptide_corr_df}
\alias{get_peptide_corr_df}
\title{Transform square correlation matrix into long data frame of correlations}
\usage{
get_peptide_corr_df(peptide_cor, peptide_annotation,
  protein_col = "ProteinName", feature_id_col = "peptide_group_label")
}
\arguments{
\item{peptide_cor}{peptide correlation matrix (square)}

\item{peptide_annotation}{long format data frame with peptide ID and their corresponding 
protein and/or gene annotations. 
See "example_peptide_annotation" for more details (to call the description, use \code{help("example_peptide_annotation")})}

\item{protein_col}{column where protein names are specified}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}
}
\value{
dataframe with correlation coefficients for each peptide pair and label
\code{same_protein} or \code{different_proteins} in column \code{same_protein}. 
Intended for use in \code{\link{plot_peptide_corr_distribution}}.
}
\description{
Transform square correlation matrix into long data frame of correlations
}
