% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{get_sample_corr_df}
\alias{get_sample_corr_df}
\title{Transform square correlation matrix into long data.frame of correlations
of the replicated samples, samples from same batch and unrelated samples}
\usage{
get_sample_corr_df(cor_proteome, sample_annotation,
  sample_id_col = "FullRunName", biospecimen_id_col = "EarTag",
  batch_col = "batch")
}
\arguments{
\item{cor_proteome}{sample correlation matrix (square)}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{biospecimen_id_col}{column in \code{sample_annotation} 
that defines a unique bio ID, which is usually a 
combination of conditions or groups.
 Tip: if such ID is absent, but can be defined from several columns,
 create new \code{biospecimen_id} column}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}
}
\value{
dataframe with the following columns, that 
are suggested to use for plotting in 
\code{\link{plot_sample_corr_distribution}} as \code{plot_param}:
\enumerate{
\item \code{replicate}
\item \code{batch_the_same}
\item \code{batch_replicate}
\item \code{batches}
}
other columns are: \enumerate{
\item \code{sample_id_1} & \code{sample_id_2}, both 
generated from \code{sample_id_col} variable
\item \code{correlation} - correlation of two corresponding samples
\item \code{batch_1} & \code{batch_2} or analogous, 
created the same as \code{sample_id_1}
}
}
\description{
useful to infer summary statistics (medians etc.) of correlation 
within/between replicates and batches.
}
