% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{gg_boxplot}
\alias{gg_boxplot}
\title{plot boxplot of data, optionally colored by batch}
\usage{
gg_boxplot(data_df_long, sample_annotation, batch_column,
  order_column = "order", measure_col = "Intensity", fill_batch = T,
  theme = "classic", title = NULL)
}
\arguments{
\item{batch_column}{}
}
\description{
plot boxplot of data, optionally colored by batch
}
