% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{handle_factor_numeric_overlap}
\alias{handle_factor_numeric_overlap}
\title{Handle factor columns that are duplicated in numeric_columns}
\usage{
handle_factor_numeric_overlap(factor_columns, numeric_columns)
}
\arguments{
\item{factor_columns}{character vector of factor columns.}

\item{numeric_columns}{character vector of numeric columns.}
}
\value{
A character vector containing \code{factor_columns} after any
  columns in \code{numeric_columns} have been removed.
}
\description{
Remove numeric columns from factor columns if overlap is detected.
}
\examples{
factor_columns <- c("batch", "injection_order")
overlap_result <- handle_factor_numeric_overlap(
    factor_columns,
    numeric_columns = "injection_order"
)
overlap_result
}
