% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{join_data_matrices}
\alias{join_data_matrices}
\title{Join data matrices}
\usage{
join_data_matrices(matrices_list, step, sample_annotation,
  measure_col = "Intensity")
}
\arguments{
\item{matrices_list}{list of matrices in \code{data_matrix} (\link{proBatch}) format to be joined}

\item{step}{normalization step (e.g. `Raw` or `Quantile_normalized` or
`qNorm_ComBat`). Useful if consecutive steps are compared in plots. Note
that in plots these are usually ordered alphabetically, so it's worth
naming with numbers, e.g. `1_raw`, `2_quantile`}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}
}
\value{
\code{df_long} (\link{proBatch}) like data frame with a row, having entries for:
\enumerate{
  \item \code{feature_id_col} (e.g. peptide name)
  \item \code{sample_id_col} (e.g. filename)
  \item \code{measure_col} (e.g. intensity/expression)
  \item \code{step} (e.g. `raw`, `quantile_norm`)
}
}
\description{
Joins 2 or more data matrices
}
\seealso{
Other matrix manipulation functions: \code{\link{convert_to_matrix}},
  \code{\link{matrix_to_long}}
}
