% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{log_transform_df}
\alias{log_transform_df}
\title{Log transformation of the data, ensuring that the row and column names
are retained}
\usage{
log_transform_df(df_long, log_base = 2, offset = 1,
  measure_col = "Intensity")
}
\arguments{
\item{df_long}{raw data matrix (features in rows and samples
in columns)}

\item{log_base}{base of the logarithm for transformation}

\item{offset}{small positive number to prevent 0 conversion to \code{-Inf}}
}
\value{
\code{df_long}-size data frame, with \code{measure_col} log transformed;
with old value in column in another column "old_intensity" if "intensity" 
was the value of \code{measure_col}
}
\description{
Log transformation of the data, ensuring that the row and column names
are retained
}
\examples{
log_transformed_matrix <- log_transform(example_proteome_matrix)

}
