% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{median_normalization}
\alias{median_normalization}
\title{Median normalization of the data}
\usage{
median_normalization(data_matrix, sample_annotation, batch_column,
  measure_column)
}
\arguments{
\item{data_matrix}{}
}
\description{
Median normalization of the data
}
