% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{merge_rare_levels}
\alias{merge_rare_levels}
\title{Replaces rare levels with other}
\usage{
merge_rare_levels(column)
}
\value{
column with rare occurrences replaced by other
}
\description{
Replaces levels with a maximal occurrence of 1 with \code{other}
}
\keyword{internal}
