% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_custom_fit}
\alias{normalize_custom_fit}
\title{normalize with the custom (continuous) fit}
\usage{
normalize_custom_fit(data_matrix, sample_annotation, batch_col, feature_id_col,
  sample_id_column, measure_col, sample_order_col, fit_func, return_long = F,
  ...)
}
\arguments{
\item{...}{other parameters, usually those of the `fit_func`}
}
\description{
normalize with the custom (continuous) fit
}
