% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalization brings the samples to the same scale}
\usage{
normalize_data(data_matrix, normalizeFunc = c("quantile",
  "medianCentering"), log_base = NULL)
}
\arguments{
\item{data_matrix}{raw data matrix (features in rows and samples
in columns)}

\item{normalizeFunc}{global batch normalization method 
(`quantile` or `MedianCentering`)}

\item{log_base}{whether to log transform data matrix 
before normalization (`NULL`, `2` or `10`)}
}
\value{
\code{data_matrix}-size matrix, with columns normalized
}
\description{
Normalization brings the samples to the same scale
}
\examples{
quantile_normalized_matrix = normalize_data(example_proteome_matrix, 
normalizeFunc = "quantile", log_base = 2)

}
