% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_global}
\alias{normalize_global}
\alias{normalize}
\title{Batch correction method allows correction of continuous sigal drift within batch and 
discrete difference across batches.}
\usage{
normalize(data_matrix, normalizeFunc = "quantile", log = NULL)
}
\arguments{
\item{data_matrix}{raw data matrix (features in rows and samples
in columns)}

\item{normalizeFunc}{global batch normalization method (`quantile` or `MedianCentering`)}

\item{log}{whether to log transform data matrix before normalization (`NULL`, `2` or `10`)}
}
\value{
`data_matrix`-size matrix, with columns normalized
}
\description{
Batch correction method allows correction of continuous sigal drift within batch and 
discrete difference across batches.
}
