% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_sample_medians}
\alias{normalize_sample_medians}
\title{Normalization by centering sample medians to global median of the data}
\usage{
normalize_sample_medians(df_long, sample_id_col = "FullRunName",
  measure_col = "Intensity")
}
\arguments{
\item{data_matrix}{log transformed long format data matrix (see `df_long`)}
}
\value{
`df_long`-size matrix, with intensity scaled to global median
}
\description{
Normalization by centering sample medians to global median of the data
}
