% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchF_operations.R
\name{pb_log_transform}
\alias{pb_log_transform}
\title{Log-transform (proBatch::log_transform_dm)}
\usage{
pb_log_transform(object, log_base = 2, offset = 1, name_suffix = "log")
}
\description{
Log-transform (proBatch::log_transform_dm)
}
