% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchF_operations.R
\name{pb_normalize}
\alias{pb_normalize}
\title{Normalize (proBatch::normalize_data_dm)}
\usage{
pb_normalize(
  object,
  method = c("quantile", "medianCentering"),
  log_base = NULL,
  offset = 1,
  name_suffix = NULL
)
}
\arguments{
\item{method}{"quantile" or "medianCentering"}
}
\description{
Normalize (proBatch::normalize_data_dm)
}
