% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{peptide_to_transition}
\alias{peptide_to_transition}
\title{separate transition column into the}
\usage{
peptide_to_transition(proteome)
}
\arguments{
\item{proteome}{}
}
\description{
separate transition column into the
}
