% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_imports.R
\name{\%>\%}
\alias{\%>\%}
\alias{foo}
\title{Convert \code{data.frame} to \code{list}.}
\usage{
foo(x)
}
\arguments{
\item{x}{A \code{data.frame} object.}
}
\description{
Convert \code{data.frame} to \code{list}.
}
\examples{
my_result <- foo(iris)

}
