% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_assessment.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Generic function for plotting per-sample mean or boxplots for initial assessment}
\usage{
plot_boxplot(x, ...)
}
\description{
Generic function for plotting per-sample mean or boxplots for initial assessment
}
