% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_clustering}
\alias{plot_clustering}
\title{cluster the data matrix to visually inspect which confounder dominates}
\usage{
plot_clustering(data_matrix, color_df, title = "Clustering of raw samples",
  distance = "euclidean", agglomeration = "complete", label_samples = T,
  label_font = 0.2, plot_title = NULL, ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. in most
function, it is assumed that this is the log transformed version of the
original data}

\item{color_df}{data frame of colors, as created by
`sample_annotation_to_colors`}

\item{title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{distance}{distance metric used for clustering}

\item{agglomeration}{agglomeration methods as used by `hclust`}

\item{...}{other parameters of `plotDendroAndColors` from `WGCNA` package}
}
\description{
cluster the data matrix to visually inspect which confounder dominates
}
\seealso{
\code{\link{hclust}}, \code{\link{sample_annotation_to_colors}},
  \code{\link{plotDendroAndColors}}
}
