% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_corr_between_samples}
\alias{plot_corr_between_samples}
\title{Plot correlation of selected samples}
\usage{
plot_corr_between_samples(data_matrix, samples_to_plot, flavor = "corrplot",
  ...)
}
\arguments{
\item{data_matrix}{features x samples matrix, with sample IDs as colnames}

\item{samples_to_plot}{string vector of samples from data_matrix}

\item{...}{parameters for the corrplot visualisation}
}
\description{
Plot correlation of selected samples
}
