% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of samples}
\usage{
plot_heatmap(data_matrix, sample_annotation = NULL,
  sample_id_col = "FullRunName", sample_annotation_col = NULL,
  sample_annotation_row = NULL, fill_the_missing = TRUE,
  cluster_rows = TRUE, cluster_cols = FALSE,
  annotation_color_list = NA,
  heatmap_color = colorRampPalette(rev(brewer.pal(n = 7, name =
  "RdYlBu")))(100), color_for_missing = "black", filename = NA,
  plot_title = NA, ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{sample_annotation}{data matrix with \enumerate{
\item  \code{sample_id_col} (this can be repeated as row names)
  \item  biological and
  \item  technical covariates (batches etc)
}; each column of sample annotation will get it's own row. 
If \code{cluster_cols = T} this will indicate,
whether sample proximity is driven by one of 
biological or technical factors}

\item{sample_id_col}{name of the column in 
sample_annotation file, where the
  filenames (colnames of the data matrix are found)}

\item{sample_annotation_col}{biological or technical 
factors to be annotated in heatmap columns}

\item{sample_annotation_row}{biological or technical 
factors to be annotated in heatmap rows}

\item{fill_the_missing}{boolean value determining if 
missing values should be
  substituted with -1 (and colored with black)}

\item{cluster_rows}{boolean value determining if rows 
should be clustered}

\item{cluster_cols}{boolean value determining if columns 
should be clustered}

\item{annotation_color_list}{list specifying colors 
for columns (samples). Best created by \code{sample_annotation_to_colors}}

\item{heatmap_color}{vector of colors used in heatmap (typicall a gradient)}

\item{color_for_missing}{special color to make missing values. 
Usually black or white, depending on \code{heatmap_color}}

\item{filename}{filepath where to save the image}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{...}{other parameters of \code{link[pheatmap]{pheatmap}}}
}
\value{
object returned by \code{link[pheatmap]{pheatmap}}
}
\description{
Plot the heatmap of samples
}
\examples{
color_scheme <- sample_annotation_to_colors (example_sample_annotation, 
factor_columns = c('MS_batch','EarTag', "Strain", 
"Diet", "digestion_batch", "Sex"),
date_columns = 'DateTime',
numeric_columns = c('order'))

heatmap_plot <- plot_heatmap(example_proteome_matrix, 
example_sample_annotation, 
sample_annotation_col = c("MS_batch",  "digestion_batch", "Diet"), 
cluster_cols = TRUE, 
annotation_color_list = color_scheme$list_of_colors,
show_rownames = FALSE, show_colnames = FALSE)

}
\seealso{
\code{\link{sample_annotation_to_colors}}, \code{\link[pheatmap]{pheatmap}}
}
