% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_iRT_trend}
\alias{plot_iRT_trend}
\title{Plot iRT measurements}
\usage{
plot_iRT_trend(df_long, sample_annotation, peptide_annotation = NULL,
  protein_col = "ProteinName", order_col = NULL, irt_pattern = "iRT",
  sample_id_col = "FullRunName", batch_col = "MS_batch.final",
  measure_col = "Intensity", feature_id_col = "peptide_group_label",
  requant = NULL, plot_title = "iRT peptide profile", ...)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{order_col}{column in \code{sample_annotation} that determines sample order. It is
used for certain diagnostics and normalisations.}

\item{irt_pattern}{substring used to identify irts proteins in the column
'ProteinName'}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{requant}{if data frame: requant values; if logical: whether to indicate
requant values (requires 'requant' or 'm_score' column in \code{df_long})}

\item{plot_title}{the string indicating the source of the peptides}

\item{...}{additional arguments to \code{\link{plot_peptide_trend}} function}
}
\value{
ggplot2 type plot of \code{measure_col} vs \code{order_col},
  faceted by \code{irt_pattern} containing proteins and (optionally) by \code{batch_col}
}
\description{
Creates a iRT facetted ggplot2 plot of the value in
\code{measure_col} vs \code{order_col} using
\code{\link{plot_peptide_trend}}. Additionally, the resulting plot can also
be facetted by batch.
}
\seealso{
Other feature-level diagnostic functions: \code{\link{plot_peptide_trend}},
  \code{\link{plot_peptides_of_one_protein}},
  \code{\link{plot_spike_ins_trend}},
  \code{\link{plot_with_fitting_curve}}
}
