% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_iRTs}
\alias{plot_iRTs}
\title{Plot iRT peptides}
\usage{
plot_iRTs(data_df_long, sample_annotation, batch_column = "MS_batch.final",
  sample_id_column = "FullRunName",
  feature_id_column = "peptide_group_label", measurement.col = "Intensity",
  order_column = "order", ...)
}
\arguments{
\item{data_df_long}{- "openSWATH" format data frame}

\item{...}{additional arguments to plot_peptide_level function}
}
\description{
Plot iRT peptides
}
