% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_peptide_corr_distribution.corrDF}
\alias{plot_peptide_corr_distribution.corrDF}
\title{Plot distribution of peptide correlation within and between proteins, 
starting from peptide correlation data frame}
\usage{
plot_peptide_corr_distribution.corrDF(corr_distribution,
  plot_title = "Correlation of peptides", theme = "classic")
}
\arguments{
\item{corr_distribution}{data frame with correlation distribution}

\item{plot_title}{title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}
}
\value{
\code{ggplot} type object with violin plot
}
\description{
Plot distribution of peptide correlation within and between proteins, 
starting from peptide correlation data frame
}
