% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_peptide_corr_distribution}
\alias{plot_peptide_corr_distribution}
\alias{plot_peptide_corr_distribution.corrDF}
\title{Create violin plot of peptide correlation distribution}
\usage{
plot_peptide_corr_distribution(
  data_matrix,
  peptide_annotation,
  protein_col = "ProteinName",
  feature_id_col = "peptide_group_label",
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = "Distribution of peptide correlation",
  theme = "classic",
  pbf_name = NULL
)

plot_peptide_corr_distribution.corrDF(
  corr_distribution,
  filename = NULL,
  width = NA,
  height = NA,
  units = c("cm", "in", "mm"),
  plot_title = "Correlation of peptides",
  theme = "classic",
  base_size = 20
)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames, or a
`ProBatchFeatures` object. When `data_matrix` is a
`ProBatchFeatures` object, `pbf_name` is used (or the current assay when
`pbf_name = NULL`).}

\item{peptide_annotation}{long format data frame with peptide ID and their
corresponding protein and/or gene annotations. When `data_matrix` is a
matrix, this argument is required. When `data_matrix` is a
`ProBatchFeatures` object and `peptide_annotation` is not provided,
rowData from the selected assay is used.}

\item{protein_col}{column where protein names are specified}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{filename}{path where the results are saved.
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only
pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image height}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overridden}

\item{pbf_name}{Assay name used when `data_matrix` is a `ProBatchFeatures`
object. If `NULL`, [pb_current_assay()] is used.}

\item{corr_distribution}{data frame with peptide correlation distribution}

\item{base_size}{base font size}
}
\value{
\code{ggplot} object (violin plot of peptide correlation)
}
\description{
Plot distribution of peptide correlations within one
protein and between proteins
}
\examples{
data(list = c("example_peptide_annotation", "example_proteome_matrix"), package = "proBatch")
peptide_corr_distribution <- plot_peptide_corr_distribution(
    example_proteome_matrix,
    example_peptide_annotation,
    protein_col = "Gene"
)

data(list = c("example_peptide_annotation", "example_proteome_matrix"), package = "proBatch")
corr_distribution <- calculate_peptide_corr_distr(
    example_proteome_matrix,
    example_peptide_annotation,
    protein_col = "Gene"
)
peptide_corr_distribution <- plot_peptide_corr_distribution.corrDF(corr_distribution)

}
\seealso{
\code{\link{calculate_peptide_corr_distr}}, \code{\link[ggplot2]{ggplot}}
}
