% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_peptide_correlation_distr_one_protein}
\alias{plot_peptide_correlation_distr_one_protein}
\title{Plot distribution curves for each step to see if they shift by data normalization}
\usage{
plot_peptide_correlation_distr_one_protein(data_matrix_list, protein_name,
  peptide_annotation, protein_col = "ProteinName",
  feature_id_col = "peptide_group_label",
  plot_title = sprintf("Distribution of peptide correlation at different correction steps,\\n                                                                            \%s protein",
  protein_name), theme = "classic")
}
\arguments{
\item{protein_name}{name of the protein, as specified in \code{protein_col}
of \code{peptide_annotation}}

\item{theme}{}
}
\description{
Plot distribution curves for each step to see if they shift by data normalization
}
