% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_peptides_of_one_protein}
\alias{plot_peptides_of_one_protein}
\title{Plot peptides of one protein}
\usage{
plot_peptides_of_one_protein(protein_name, peptide_annotation = NULL,
  protein_col = "ProteinName", df_long, sample_annotation,
  sample_id_col = "FullRunName", measure_col = "Intensity",
  feature_id_col = "peptide_group_label", geom = c("point", "line"),
  color_by_col = NULL, color_by_value = NULL, batch_col = "MS_batch",
  color_by_batch = FALSE, color_scheme = "brewer",
  order_col = "order", vline_color = "red", facet_col = FALSE,
  plot_title = sprintf("Peptides of \%s protein", protein_name),
  theme = "classic", ...)
}
\arguments{
\item{protein_name}{name of the protein as defined in \code{ProteinName}}

\item{peptide_annotation}{long format data frame with peptide ID and their corresponding 
protein and/or gene annotations (analogous to )}

\item{protein_col}{column where protein names are specified}

\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{geom}{whether to show the feature as points and/or connect by lines}

\item{color_by_col}{column to color point by certain value denoted 
by \code{color_by_value}. Design with inferred/requant values in openSWATH output data, 
which means argument value has to be set to `m_score`.}

\item{color_by_value}{value in \code{color_by_col} to color. For OpenSWATH data,
this argument value has to be set to `2` (this is an `m_score` value for requants).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{color_by_batch}{(logical) whether to color points and connecting lines by batch}

\item{color_scheme}{color scheme for \code{ggplot} representation of batches}

\item{order_col}{column in \code{sample_annotation} that determines sample order. It is
used for certain diagnostics and normalisations.}

\item{vline_color}{color of vertical lines, typically separating 
different MS batches in ordered runs; 
should be `NULL` for experiments without intrinsic order.}

\item{facet_col}{column  in `sample_annotation` with a batch factor to separate 
plots into facets; usually 2nd to `batch_col`. Most meaningful for multi-instrument 
MS experiments (where each instrument has its own order-associated effects) 
or simultaneous examination of two batch factors (e.g. preparation day and measurement day)}

\item{plot_title}{the string indicating the source of the peptides}

\item{theme}{plot theme (default is 'classical'; other options not
implemented)}

\item{...}{additional arguments to \code{\link{plot_single_feature}} function}
}
\value{
ggplot2 type plot of \code{measure_col} vs \code{order_col},
  faceted by \code{spike_ins} containing proteins and (optionally) by \code{batch_col}
}
\description{
Creates a spike-in facetted ggplot2 plot of the value in
\code{measure_col} vs \code{order_col} using
\code{\link{plot_single_feature}}. Additionally, the resulting plot can also
be facetted by batch.
}
\examples{
plot_peptides_of_one_protein (protein_name = "Haao",  
protein_col = "Gene", df_long = example_proteome, 
example_sample_annotation, 
order_col = 'order', sample_id_col = 'FullRunName', 
batch_col = 'MS_batch')

}
\seealso{
Other feature-level diagnostic functions: \code{\link{plot_iRT}},
  \code{\link{plot_single_feature}},
  \code{\link{plot_spike_in}},
  \code{\link{plot_with_fitting_curve}}
}
\concept{feature-level diagnostic functions}
