% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_prot_corr_distribution}
\alias{plot_prot_corr_distribution}
\title{Plot distribution of peptide correlations within one protein and between proteins}
\usage{
plot_prot_corr_distribution(data_matrix, peptide_annotation,
  protein_col = "ProteinName", feature_id_col = "peptide_group_label",
  plot_title = "Distribution of peptide correlation", theme = "classic")
}
\arguments{
\item{theme}{}
}
\description{
Plot distribution of peptide correlations within one protein and between proteins
}
