% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_sample_corr_distribution.corrDF}
\alias{plot_sample_corr_distribution.corrDF}
\title{Plot correlation distribution, starting from the correlation matrix}
\usage{
plot_sample_corr_distribution.corrDF(corr_distribution,
  plot_title = "Sample correlation distribution",
  plot_param = "batch_replicate", theme = "classic")
}
\arguments{
\item{corr_distribution}{data frame with correlation distribution}

\item{plot_title}{title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{plot_param}{one of the columns, as returned by \code{get_sample_corr_df}}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}
}
\value{
\code{ggplot} type object with violin plot
}
\description{
Plot correlation distribution, starting from the correlation matrix
}
