% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_samples_corrplot}
\alias{plot_samples_corrplot}
\title{Sample correlation matrix (heatmap)}
\usage{
plot_samples_corrplot(data_matrix, samples_to_plot = NULL,
  flavor = "corrplot", filename = NULL, width = NA, height = NA,
  unit = c("cm", "in", "mm"),
  plot_title = "Correlation matrix of samples", ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{samples_to_plot}{string vector of samples in \code{data_matrix} to be used in the plot}

\item{filename}{path where the results are saved. If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{...}{parameters for the \code{\link[corrplot]{corrplot.mixed}} or
\code{\link[pheatmap]{pheatmap}} visualisation, for details see examples and
  help to corresponding functions}
}
\description{
Plot correlation of selected samples
}
\examples{
### Example 1: Plot heatmap of pre-specified samples
#'\dontrun{specified_samples = sample_annotation \%>\%
filter(RunID \%in\% paste('Run', 110:115, sep = '')) \%>\%
pull(FullRunName)
plot_samples_corr_heatmap(data_matrix, sample_to_plot = specified samples,
 flavor = 'pheatmap', cluster_rows = F, cluster_cols = F)

}
### Example 2: Plot corrplot of pre-specified samples
#' #'\dontrun{specified_samples = sample_annotation \%>\%
filter(RunID \%in\% paste('Run', 110:115, sep = '')) \%>\%
pull(FullRunName)
plot_samples_corr_heatmap(data_matrix, sample_to_plot = specified samples,
 flavor = 'corrplot', lower = "ellipse", upper = "number",
 tl.col = "black", diag = 'l', tl.pos = "lt", number.cex=0.75, tl.cex = .75)

}
}
\seealso{
\code{\link[pheatmap]{pheatmap}}, \code{\link[corrplot]{corrplot.mixed}}
}
