% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_spike_ins}
\alias{plot_spike_ins}
\title{Plot Spike-in peptides/proteins}
\usage{
plot_spike_ins(data_df_long, sample_annotation, spike_ins = "BOVIN",
  order_column = "order", measurement.col = "Intensity",
  batch_column = "MS_batch", feature_id_column = "peptide_group_label", ...)
}
\arguments{
\item{...}{}
}
\description{
Plot Spike-in peptides/proteins
}
