% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_proteome.R
\name{remove_peptides_with_missing_batch}
\alias{remove_peptides_with_missing_batch}
\title{Remove features missing in at least one batch}
\usage{
remove_peptides_with_missing_batch(df_long, sample_annotation,
  batch_col = "MS_batch.final", feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName")
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}
}
\value{
\code{df_long} (\link{proBatch}) like data frame freed of features that were not detected in each batch
}
\description{
Cleans dataset \code{df_long} (\link{proBatch}) by removing all features that are not present in every batch
}
\details{
useful for some downstream functions as ComBat normalization, that
  would not work otherwise
}
\seealso{
Other dataset cleaning functions: \code{\link{clean_requants}},
  \code{\link{summarize_peptides}}
}
\concept{dataset cleaning functions}
