% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{sample_annotation_to_colors}
\alias{sample_annotation_to_colors}
\alias{sample_annotation_to_colors.default}
\alias{sample_annotation_to_colors.ProBatchFeatures}
\title{Generate colors for sample annotation}
\usage{
\method{sample_annotation_to_colors}{default}(
  sample_annotation,
  sample_id_col = "FullRunName",
  factor_columns = NULL,
  numeric_columns = NULL,
  rare_categories_to_other = TRUE,
  guess_factors = FALSE,
  numeric_palette_type = "brewer",
  ...
)

\method{sample_annotation_to_colors}{ProBatchFeatures}(sample_annotation, ...)

sample_annotation_to_colors(sample_annotation, ...)
}
\arguments{
\item{sample_annotation}{Input object supplied to the generic (data
frame or `ProBatchFeatures`).}

\item{sample_id_col}{name of the column in \code{sample_annotation} table,
where the filenames (colnames of the \code{data_matrix} are found).}

\item{factor_columns}{columns of \code{sample_annotation} to be
treated as factors. Sometimes categorical variables are depicted as integers
(e.g. in column "Batch", values are 1, 2 and 3), specification here allows
to map them correctly to qualitative palettes.}

\item{numeric_columns}{columns of \code{sample_annotation} to be
treated as continuous numeric values.}

\item{rare_categories_to_other}{if \code{True} rare categories
will be merged into the value \code{"other"}}

\item{guess_factors}{whether attempt which of the \code{factor_columns}
are actually numeric}

\item{numeric_palette_type}{palette to be used for
numeric values coloring (can be \code{'brewer' and 'viridis'})}

\item{...}{Additional arguments forwarded to method implementations.}
}
\value{
list of colors for the selected annotation columns. Use
\code{\link{color_list_to_df}} if a data frame representation is needed.
}
\description{
Convert the sample annotation data frame to list of colors the list is named
as columns included to use in plotting functions
}
\examples{
data("example_sample_annotation", package = "proBatch")
color_scheme <- sample_annotation_to_colors(
    example_sample_annotation,
    factor_columns = c(
        "MS_batch", "EarTag", "Strain",
        "Diet", "digestion_batch", "Sex"
    ),
    numeric_columns = c("DateTime", "order")
)
}
