% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_color_scheme}
\alias{sample_color_scheme}
\title{Sample color annotation}
\format{An object of class \code{list} of length 3.}
\usage{
sample_color_scheme
}
\description{
This is an color scheme generated from example sample annotation
}
\keyword{datasets}
