% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{subset_keep_cols}
\alias{subset_keep_cols}
\title{Subset columns according to `keep_all` argument}
\usage{
subset_keep_cols(
  df,
  keep_all = "default",
  default_cols = names(df),
  minimal_cols = default_cols
)
}
\arguments{
\item{df}{data frame to subset}

\item{keep_all}{one of `"all"`, `"default"` or `"minimal"`}

\item{default_cols}{columns to keep when `keep_all = "default"`}

\item{minimal_cols}{columns to keep when `keep_all = "minimal"`}
}
\value{
data frame with selected columns
}
\description{
Helper for data transformation functions to consistently keep a
predefined set of columns.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9)
subset_keep_cols(df, keep_all = "minimal", minimal_cols = c("a", "c"))
}
\keyword{internal}
