% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{summarize_peptides}
\alias{summarize_peptides}
\title{convert sample annotation matrix to color data frame}
\usage{
summarize_peptides(proteome)
}
\arguments{
\item{proteome}{}
}
\description{
convert sample annotation matrix to color data frame
}
