% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_calculation.R
\name{calculate_feature_CV}
\alias{calculate_feature_CV}
\title{Calculate CV distribution for each feature}
\usage{
calculate_feature_CV(df_long, sample_annotation = NULL,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName", measure_col = "Intensity",
  batch_col = NULL)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} 
and a \code{measure_col}, but usually also an \code{m_score} (in OpenSWATH 
output result file). See \code{help("example_proteome")} for more details.}

\item{sample_annotation}{}

\item{feature_id_col}{}

\item{sample_id_col}{}

\item{measure_col}{}

\item{batch_col}{}
}
\value{
data frame with Total CV for each feature & (optionally) per-batch CV
}
\description{
Calculate CV distribution for each feature
}
\examples{
CV_df = calculate_feature_CV(example_proteome, 
sample_annotation = sample_annotation_AgingMice, 
measure_col = 'Intensity', 
batch_col = 'MS_batch')
}
