% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{calculate_PVCA}
\alias{calculate_PVCA}
\alias{calculate_PVCA.default}
\alias{calculate_PVCA.ProBatchFeatures}
\title{Calculate variance distribution by variable}
\usage{
\method{calculate_PVCA}{default}(
  data_matrix,
  sample_annotation,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName",
  factors_for_PVCA = c("MS_batch", "digestion_batch", "Diet", "Sex", "Strain"),
  pca_threshold = 0.6,
  variance_threshold = 0.01,
  fill_the_missing = -1,
  ...
)

\method{calculate_PVCA}{ProBatchFeatures}(
  data_matrix,
  pbf_name = NULL,
  sample_annotation = NULL,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName",
  ...
)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. See
"example_proteome_matrix" for more details (to call the description, use
\code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with:
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names)
  \item biological covariates
  \item technical covariates (batches etc) }.
  See \code{help("example_sample_annotation")}}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this
corresponds to the row names.}

\item{sample_id_col}{name of the column in \code{sample_annotation} table,
where the filenames (colnames of the \code{data_matrix} are found).}

\item{factors_for_PVCA}{vector of factors from \code{sample_annotation}, that
are used in PVCA analysis}

\item{pca_threshold}{the percentile value of the minimum amount of the
variabilities that the selected principal components need to explain}

\item{variance_threshold}{the percentile value of weight each of the factors
needs to explain (the rest will be lumped together)}

\item{fill_the_missing}{numeric value determining how  missing values
should be substituted. If \code{NULL}, features with
missing values are excluded.}

\item{...}{Additional arguments forwarded between methods.}

\item{pbf_name}{Assay name(s) used when \code{data_matrix}
is a \code{ProBatchFeatures}.}
}
\value{
data frame of weights of Principal Variance Components
}
\description{
Calculate variance distribution by variable
}
\examples{
data(
    list = c("example_proteome_matrix", "example_sample_annotation"),
    package = "proBatch")
matrix_test <- na.omit(example_proteome_matrix)[1:50, ]
pvca_df <- calculate_PVCA(matrix_test, example_sample_annotation,
    factors_for_PVCA = c(
        "MS_batch", "digestion_batch", "Diet", "Sex", "Strain"),
    pca_threshold = .6, variance_threshold = .01, fill_the_missing = -1
)
}
