% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{calculate_PVCA}
\alias{calculate_PVCA}
\title{Calculate variance distribution by variable}
\usage{
calculate_PVCA(data_matrix, sample_annotation, factors_for_PVCA,
  pca_threshold, variance_threshold = Inf)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{factors_for_PVCA}{vector of factors from \code{sample_annotation}, that
are used in PVCA analysis}

\item{pca_threshold}{the percentile value of the minimum amount of the
variabilities that the selected principal components need to explain}

\item{variance_threshold}{the percentile value of weight each of the covariates
needs to explain (the rest will be lumped together)}
}
\description{
Calculate variance distribution by variable
}
