% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effects.R
\name{center_feature_batch_medians_dm}
\alias{center_feature_batch_medians_dm}
\title{DEPRECATED: center_feature_batch_medians_dm}
\usage{
center_feature_batch_medians_dm(
  data_matrix,
  sample_annotation,
  sample_id_col = "FullRunName",
  batch_col = "MS_batch",
  feature_id_col = "peptide_group_label",
  measure_col = "Intensity"
)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. See
"example_proteome_matrix" for more details (to call the description, use
\code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with:
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names)
  \item biological covariates
  \item technical covariates (batches etc) }.
  See \code{help("example_sample_annotation")}}

\item{sample_id_col}{name of the column in \code{sample_annotation} table,
where the filenames (colnames of the \code{data_matrix} are found).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison (or other, non-batch factor to be mapped to color in plots).}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this
corresponds to the row names.}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}
}
\value{
A numeric matrix of batch-corrected values
(features \eqn{\times} samples).
}
\description{
Use \code{\link[=center_feature_batch]{center_feature_batch()}} with
\verb{format="wide", stat="medians"}.
}
\examples{
# Use the unified replacement with `format = "wide"` and `stat = "medians"`.
args(center_feature_batch)
}
