% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_proteome.R
\name{clean_proteome}
\alias{clean_proteome}
\alias{remove_peptides_with_missing_batch}
\alias{clean_proteome}
\alias{summarize_peptides}
\title{Remove peptides that are missing in the whole batch}
\usage{
remove_peptides_with_missing_batch(df_long, batch_column = "MS_batch.final",
  feature_id_column = "peptide_group_label")

summarize_peptides(df_long, sample_id_col = "FullRunName",
  feature_id_column = "peptide_group_label")
}
\value{
`df_long`-like data frame free of peptides that were not detected in each batch
}
\description{
Remove peptides that are missing in the whole batch

summarize peptides by sample (ranking) and on the contrary, across peptide-wise across samples
}
\details{
useful for some downstream functions as ComBat normalization, that would not work otherwise
}
