% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_proteome.R
\name{clean_requants}
\alias{clean_requants}
\title{Remove requanted and sparse features}
\usage{
clean_requants(df_long, sample_annotation, batch_col = "MS_batch.final",
  feature_id_col = "peptide_group_label", threshold_batch = 0.3,
  threshold_global = 0.3)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{threshold_batch}{maximally tollerated fraction of missing values for a
feature in a batch}

\item{threshold_global}{maximally tollerated fraction of globally missing
values for a feature}
}
\value{
\code{df_long} (\link{proBatch}) like data frame filtered as follow:
  \itemize{ \item remove requant values \item remove features not meeting
  batch or global sparsness thresholds }
}
\description{
Cleans dataset \code{df_long} (\link{proBatch}) by removing requanted features and features not
meeting user define sparsness criterias.
}
\seealso{
Other dataset cleaning functions: \code{\link{remove_peptides_with_missing_batch}},
  \code{\link{summarize_peptides}}
}
