% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{color_list_to_df}
\alias{color_list_to_df}
\title{Color list to data frame}
\usage{
color_list_to_df(color_list, sample_annotation)
}
\arguments{
\item{color_list}{list of colors}

\item{sample_annotation}{factor-based configuration of the sample annotation}
}
\value{
a data frame representation of the input color list
}
\description{
Turn color list to df (some plotting functions require the latter)
}
