% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{color_list_to_df}
\alias{color_list_to_df}
\title{Turn color list to df (some plotting functions require the latter)}
\usage{
color_list_to_df(color_list, sample_annotation)
}
\arguments{
\item{sample_annotation}{}
}
\description{
Turn color list to df (some plotting functions require the latter)
}
