% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{convert_annotation_classes}
\alias{convert_annotation_classes}
\title{Convert factor and numeric columns}
\usage{
convert_annotation_classes(df, factor_columns, numeric_columns)
}
\arguments{
\item{df}{data frame with sample annotations.}

\item{factor_columns}{character vector of factor columns.}

\item{numeric_columns}{character vector of numeric columns.}
}
\value{
data frame with converted columns.
}
\description{
Convert specified factor columns to factor type and numeric columns to
numeric.
}
