% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effects.R
\name{correct_with_ComBat}
\alias{correct_with_ComBat}
\title{Adjusts for discrete batch effects using ComBat}
\usage{
correct_with_ComBat(data_matrix, sample_annotation,
  sample_id_col = "FullRunName", batch_col = "MS_batch",
  par.prior = TRUE)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{sample_annotation}{data frame with sample ID, technical (e.g. MS batches) 
and biological (e.g. Diet) covariates}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{batch_col}{column in \code{sample_annotation} that should be 
used for batch comparison}

\item{par.prior}{whether parametrical or non-parametrical prior should be used}
}
\value{
\code{data_matrix}-size data matrix with batch-effect corrected by
  \code{ComBat}
}
\description{
Standardized input-output ComBat normalization ComBat allows users to adjust
for batch effects in datasets where the batch covariate is known, using
methodology described in Johnson et al. 2007. It uses either parametric or
non-parametric empirical Bayes frameworks for adjusting data for batch
effects. Users are returned an expression matrix that has been corrected for
batch effects. The input data are assumed to be cleaned and normalized before
batch effect removal.
}
\examples{
correct_with_ComBat(example_proteome_matrix, example_sample_annotation)

}
