% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{correct_with_ComBat}
\alias{correct_with_ComBat}
\title{Standardized input-output ComBat normalization}
\usage{
correct_with_ComBat(data_matrix, sample_annotation,
  batch_column = "MS_batch.final", par.prior = TRUE)
}
\arguments{
\item{data_matrix}{}
}
\description{
Standardized input-output ComBat normalization
}
