% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effects.R
\name{correct_with_removeBatchEffect_dm}
\alias{correct_with_removeBatchEffect_dm}
\title{Batch effect correction with removeBatchEffect from limma}
\usage{
correct_with_removeBatchEffect_dm(
  data_matrix,
  sample_annotation,
  feature_id_col = "peptide_group_label",
  measure_col = "Intensity",
  sample_id_col = "FullRunName",
  batch_col = "MS_batch",
  covariates_cols = NULL,
  fill_the_missing = NULL,
  ...
)

correct_with_removeBatchEffect_dm(
  data_matrix,
  sample_annotation,
  feature_id_col = "peptide_group_label",
  measure_col = "Intensity",
  sample_id_col = "FullRunName",
  batch_col = "MS_batch",
  covariates_cols = NULL,
  fill_the_missing = NULL,
  ...
)
}
\arguments{
\item{data_matrix}{data matrix with features in rows and samples in columns}

\item{sample_annotation}{data frame with sample annotations}

\item{feature_id_col}{column name in \code{data_matrix} with feature IDs}

\item{measure_col}{column name in \code{data_matrix} with measured values}

\item{sample_id_col}{column name in \code{sample_annotation} with sample IDs}

\item{batch_col}{column name in \code{sample_annotation} with batch IDs}

\item{covariates_cols}{vector of column names in \code{sample_annotation}
with covariates to include in the model}

\item{fill_the_missing}{numeric value used to impute missing measurements
before correction. If \code{FALSE} rows with missing values are removed.}

\item{...}{other parameters to pass to \code{removeBatchEffect}}
}
\value{
data matrix with batch effects removed
}
\description{
Batch effect correction with removeBatchEffect.
}
\section{Functions}{
\itemize{
\item \code{correct_with_removeBatchEffect_dm()}: Deprecated wrapper for
`correct_with_removeBatchEffect(format = "wide")`.

}}
\examples{
data(
    list = c("example_sample_annotation", "example_proteome_matrix"),
    package = "proBatch"
)
example_proteome_small <- example_proteome_matrix[1:100, ]
batch_corrected_matrix <- correct_with_removeBatchEffect_dm(
    example_proteome_small,
    example_sample_annotation,
    batch_col = "MS_batch",
    covariates_cols = c("Diet", "Sex")
)
}
\seealso{
\code{\link[limma:removeBatchEffect]{removeBatchEffect}}
}
