% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{create_peptide_annotation}
\alias{create_peptide_annotation}
\title{Create light-weight peptide annotation data frame for selection of illustrative proteins}
\usage{
create_peptide_annotation(df_long, peptide_col = "peptide_group_label",
  annotation_col = c("RT", "Intensity", "ProteinName"))
}
\arguments{
\item{peptide_col}{column containing peptide ID}

\item{protein_col}{one or more columns contatining protein ID}
}
\description{
Create light-weight peptide annotation data frame for selection of illustrative proteins
}
\examples{
\\donotrun{
peptide_annotation = create_peptide_annotation(example_proteome)
peptide_summary =
}

}
\seealso{
\code{\link{plot_peptides_of_one_protein}}, \code{\link{plot_corrplot_protein}},
\code{\link{plot_within_prot_distribution}}
}
