% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{date_to_sample_order}
\alias{date_to_sample_order}
\title{Convert date/time to POSIXct and rank samples by it}
\usage{
date_to_sample_order(sample_annotation, time_column = c("RunDate",
  "RunTime"), new_time_column = "DateTime",
  dateTimeFormat = c("\%b_\%d", "\%H:\%M:\%S"),
  new_order_col = "order", instrument_col = "instrument")
}
\arguments{
\item{sample_annotation}{data frame with: 
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names) 
  \item biological covariates 
  \item technical covariates (batches etc) }. 
  See \code{help("example_sample_annotation")}}

\item{time_column}{name of the column(s) where run date & time are specified.
These will be used to determine the run order}

\item{new_time_column}{name of the new column to which date&time will be
converted to}

\item{dateTimeFormat}{POSIX format of the date and time. 
See \code{\link{as.POSIXct}} from base R for details}

\item{new_order_col}{name of column with generated the order of sample run 
based on time columns}

\item{instrument_col}{column, denoting different 
instrument used for measurements}
}
\value{
sample annotation file with a new column \code{new_time_column} with
  POSIX-formatted date & \code{new_order_col} used 
  in some diagnostic plots (e.g.
  \code{\link{plot_iRT}}, \code{\link{plot_sample_mean}})
}
\description{
Converts date/time columns fo sample_annotation to POSIXct format and
calculates sample run rank in order column
}
\examples{
sample_annotation_wOrder <- date_to_sample_order(
example_sample_annotation,
time_column = c('RunDate','RunTime'),
new_time_column = 'new_DateTime',
dateTimeFormat = c("\%b_\%d", "\%H:\%M:\%S"),
new_order_col = 'new_order',
instrument_col = NULL)

}
