% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{date_to_sample_order}
\alias{date_to_sample_order}
\title{convert date to order}
\usage{
date_to_sample_order(sample_annotation, time_column,
  new_time_column = "DateTime", dateTimeFormat = c("\%b_\%d",
  "\%H:\%M:\%S"), order_column = "order")
}
\arguments{
\item{order_column}{}
}
\description{
convert date to order
}
