% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{date_to_sample_order}
\alias{date_to_sample_order}
\title{Convert date/time to POSIXct and rank samples by it}
\usage{
date_to_sample_order(sample_annotation, time_column = c("RunDate", "RunTime"),
  new_time_column = "DateTime", dateTimeFormat = c("\%b_\%d",
  "\%H:\%M:\%S"), order_column = "order", instrument_col = "instrument")
}
\arguments{
\item{sample_annotation}{data matrix with: \enumerate{ \item `sample_id_col`
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{time_column}{name of the column(s) where run date & time are specified.
These will be used to determine the run order}

\item{new_time_column}{name of the new column to which date&time will be
converted to}

\item{dateTimeFormat}{POSIX format of the date and time. See \code{\link{as.POSIXct}}
from base R for details}

\item{order_column}{name of the new column that determines sample order. Will
be used for certain diagnostics and normalisations}
}
\value{
sample annotation file with column names as 'new_time_column' with
  POSIX-formatted date & `order_column` used in some diagnostic plots (e.g.
  \code{\link{plot_iRTs}}, \code{\link{plot_sample_mean}})
}
\description{
Converts date/time columns fo sample_annotation to POSIXct format and
calculates sample run rank in order column
}
