% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{dates_to_posix}
\alias{dates_to_posix}
\title{convert date/time column to POSIX format
required to keep number-like behaviour}
\usage{
dates_to_posix(sample_annotation, time_column, new_time_column = NULL,
  dateTimeFormat = c("\%b_\%d", "\%H:\%M:\%S"))
}
\arguments{
\item{dateTimeFormat}{}
}
\description{
convert date/time column to POSIX format
required to keep number-like behaviour
}
