% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{define_batches_by_MS_pauses}
\alias{define_batches_by_MS_pauses}
\title{convert by batch}
\usage{
define_batches_by_MS_pauses(sample_annotation, threshold,
  runtime_col = "RunDateTime", minimal_batch_size = 5,
  instrument_col = "instr", batch_name = "MS_batch")
}
\arguments{
\item{batch_name}{}
}
\description{
convert by batch
}
