% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_proteome.R
\name{define_batches_by_MS_pauses}
\alias{define_batches_by_MS_pauses}
\title{Identify stretches of time between runs that are long and split a batches by them}
\usage{
define_batches_by_MS_pauses(date_vector, threshold, minimal_batch_size = 5,
  batch_name = "MS_batch")
}
\arguments{
\item{date_vector}{POSIX or numeric-like vector corresponding to the sample
MS profile acquisition timepoint}

\item{threshold}{time difference that would mean there was an interruption}

\item{minimal_batch_size}{minimal number of samples in a batch}

\item{batch_name}{string with a self-explanatory name for the batch (e.g. `MS_batch` for MS-proteomics)}
}
\value{
vector of batches for each sample
}
\description{
Identify stretches of time between runs that are long and split a batches by them
}
