% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{define_sample_order}
\alias{define_sample_order}
\title{Defining sample order internally}
\usage{
define_sample_order(order_col, sample_annotation, facet_col, batch_col,
  df_long, sample_id_col, color_by_batch)
}
\arguments{
\item{order_col}{}

\item{sample_annotation}{}

\item{facet_col}{}

\item{batch_col}{}

\item{df_long}{}

\item{sample_id_col}{}

\item{color_by_batch}{}
}
\description{
Defining sample order internally
}
\seealso{
\link{plot_sample_mean_or_boxplot}, \link{feature_level_diagnostics}
}
