% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_peptide_annotation}
\alias{example_peptide_annotation}
\title{Peptide annotation data}
\format{A data frame with 535 rows and 10 variables:
\describe{
  \item{peptide_group_label}{peptide group label ID, identical to 
  \code{peptide_group_label} in \code{example_proteome}}
  \item{Gene}{HUGO gene ID}
  \item{ProteinName}{protein group name as specified in 
  \code{example_proteome}}
  }}
\usage{
example_peptide_annotation
}
\description{
This is data from Aging study annotated with gene names
}
\keyword{datasets}
