% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_peptide_annotation}
\alias{example_peptide_annotation}
\title{Peptide annotation data}
\format{A data frame with 200625 rows and 8 variables:
\describe{
  \item{Peptide}{peptide group label ID, identical to `peptide_group_label` in `example_proteome`}
  \item{Gene}{HUGO gene ID}
  \item{ProteinName}{protein group name as specified in `example_proteome`}
  ... other parameters determined by 'summarize_peptides' function
  }}
\usage{
example_peptide_annotation
}
\description{
This is data from Evan's aging study annotated with gene names
}
\keyword{datasets}
