% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_proteome}
\alias{example_proteome}
\title{Example protein data in long format}
\format{An object of class \code{data.frame} with 126053 rows and 6 columns.}
\source{
PRIDE ID will be added upon the publication of the dataset
}
\usage{
example_proteome
}
\description{
This is OpenSWATH-output data from Aging study with all iRT, spike-in peptides,
few representative peptides and proteins for signal improvement demonstration.
Using \code{matrix_to_long} can be converted to \code{example_proteome_matrix}
}
\keyword{datasets}
