% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_proteome}
\alias{example_proteome}
\title{Example protein data in long format}
\format{A data frame with 124655 rows and 5 variables:
\describe{
  \item{peptide_group_label}{peptide ID, which is regular feature level. 
  This column is mostly used as \code{feature_id_col}}
  \item{Intensity}{peptide group intensity in given sample. 
  Used in function as \code{measure_col}}
  \item{ProteinName}{Protein group ID, specified as 
  N/UniProtID1|UniProtID2|...,
  where N is number of protein peptide group maps to. If 
  1/UniProtID, then this is proteotypic peptide}
  \item{Gene}{conventional gene name of corresponding ProteinName}
  \item{FullRunName}{name of the file, in most functions used for \code{sample_id_col}}
...
}}
\source{
PRIDE ID will be added in future
}
\usage{
example_proteome
}
\description{
This is data from Aging study with all iRT, spike-in peptides,
few random peptides and QTL proteins for biological 
signal improvement demonstration
}
\keyword{datasets}
