% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_proteome_matrix}
\alias{example_proteome_matrix}
\title{Example protein data in matrix}
\format{A matrix with 534 rows and 233 columns:}
\source{
PRIDE ID will be added in future
}
\usage{
example_proteome_matrix
}
\description{
This is measurement data from Aging study with columns
representing samples and rows representing peptides
}
\keyword{datasets}
