% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_proteome_matrix}
\alias{example_proteome_matrix}
\title{Example protein data in matrix}
\format{A matrix with 535 rows and 233 columns:}
\source{
PRIDE ID will be added upon the publication of the dataset
}
\usage{
example_proteome_matrix
}
\description{
This is measurement data from Aging study with columns
representing samples and rows representing peptides. Generated by 
\code{long_to_matrix}
}
\keyword{datasets}
