% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sample_annotation}
\alias{example_sample_annotation}
\title{Sample annotation data version 1}
\format{A data frame with 233 rows and 11 variables:
\describe{
  \item{FullRunName}{name of the file, in most functions used for \code{sample_id_col}}
  \item{MS_batch}{mass-spectrometry batch: 7-level factor of 
  manually annotated batches}
  \item{EarTag}{mouse ID, i.e. ID of the biological object}
  \item{Strain}{mouse strain ID - biological covariate #1}
  \item{Diet}{diet - either \code{HFD} = `High Fat Diet` or \code{CD} = `Chow Diet`.
  \code{Mix} stands for mixture of several samples}
  \item{Sex}{mice sex - 3-level biological covariate. Possible values - ''}
  \item{RunDate}{mass-spectrometry running date. In combination 
  with \code{RunTime} used for running order determination}
  \item{RunTime}{mass-spectrometry running time. In combination 
  with \code{RunDate} used for running order determination}
  \item{DateTime}{numeric date and time generated by \code{date_to_sample_order}}
  \item{order}{order of samples generated by sorting DateTime 
  in \code{date_to_sample_order}}
  \item{digestion_batch}{peptide digestion batch: 5-level 
  factor of manually annotated batches}
  ...
}}
\usage{
example_sample_annotation
}
\description{
This is data from BXD aging study with mock instruments to show how
instrument-specific functionality works
}
\keyword{datasets}
