% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sample_annotation}
\alias{example_sample_annotation}
\title{Sample annotation data version 1}
\format{A data frame with 375 rows and 14 variables:
\describe{
  \item{FullRunName}{name of the file, in most functions used for `sample_id_col`}
  \item{MS_batch}{mass-spectrometry batch: 7-level factor of manually annotated batches}
  \item{EarTag}{mouse ID, i.e. ID of the biological object}
  \item{Strain}{mouse strain ID - biological covariate #1}
  \item{Diet}{diet - either `HFD` = `High Fat Diet` or `CD` = `Chow Diet`.
  `Mix` stands for mixture of several samples}
  \item{Sex}{mice sex - 3-level biological covariate. Possible values - ''}
  \item{Age_Days}{mice age at sampling - numeric biological covariate}
  \item{RunDate}{mass-spectrometry running date. In combination with `RunTime` used for running order determination}
  \item{RunTime}{mass-spectrometry running time. In combination with `RunDate` used for running order determination}
  \item{SacrificeDate}{date of mouse sacrifice - technical covariate}
  \item{ProteinPrepDate}{date of protein preparation - technical covariate}
  \item{DateTime}{numeric date and time generated by \code{date_to_sample_order}}
  \item{order}{order of samples generated by sorting DateTime in \code{date_to_sample_order}}
  \item{instrument}{different instrument used for data acquisition}
  ...
}}
\usage{
example_sample_annotation
}
\description{
This is data from BXD aging study with mock instruments to show how
instrument-specific functionality works
}
\keyword{datasets}
