% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_non_linear.R
\name{fit_nonlinear}
\alias{fit_nonlinear}
\title{Fit a non-linear trend}
\usage{
fit_nonlinear(dataDF, batch.size, response.var = "y", expl.var = "x",
  noFitRequants = F, fitFunc = "loess_regression", with_df = F,
  loess.span = 0.75, abs.threshold = 5, pct.threshold = 0.2, ...)
}
\arguments{
\item{batch.size}{the total number of samples in the batch to compute for percentage threshold}

\item{response.var}{the name of the column in dataDF with the response variable}

\item{expl.var}{the name of the column in dataDF with the explanatory variable}

\item{noFitRequants}{(logical) whether to fit requanted values}

\item{fitFunc}{function to use for the fit (`loess_regression`)}

\item{loess.span}{the parameter α which controls the degree of smoothing for loess}

\item{abs.threshold}{the absolute threshold to filter data for curve fitting}

\item{pct.threshold}{the percentage threshold to filter data for curve fitting}

\item{...}{additional paramters to be passed to the fitting function}
}
\value{
vector of fitted response values
}
\description{
Fit a non-linear trend
}
\keyword{internal}
