% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_non_linear.R
\name{fit_nonlinear}
\alias{fit_nonlinear}
\title{Fit a non-linear trend}
\usage{
fit_nonlinear(dataDF, response.var = "y", expl.var = "x",
  noFitRequants = F, fitFunc = "kernel_smooth", with_df = F, ...)
}
\arguments{
\item{response.var}{the name of the column in dataDF with the response variable}

\item{expl.var}{the name of the column in dataDF with the explanatory variable}

\item{noFitRequants}{(logical) whether to fit requanted values}

\item{fitFunc}{function to use for the fit (`kernel_smooth`, `smooth_spline`, or `loess_regression`)}

\item{...}{additional paramters to be passed to the fitting function}
}
\value{
vector of fitted response values
}
\description{
Fit a non-linear trend
}
\keyword{internal}
