% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{generate_colors_for_numeric}
\alias{generate_colors_for_numeric}
\title{generates color list}
\usage{
generate_colors_for_numeric(num_col, palette_type = "brewer", i = 1,
  granularity = 10)
}
\arguments{
\item{num_col}{a vector of type numeric of factor to generate colors for}

\item{palette_type}{'brewer' or 'viridis'}

\item{i}{if \code{palette_type} is 'brewer' the palette argument to \code{brewer_pal}. If \code{palette_type} is 'viridis' the option argument to virids_pal ()}

\item{granularity}{the breaks to use when generating colors for num_col}
}
\value{

}
\description{
generates a list of colors for the dataframe with all columns numeric (or date)
}
