% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{generate_colors_for_numeric}
\alias{generate_colors_for_numeric}
\title{generate a list of colors for the dataframe with all columns numeric (or date)}
\usage{
generate_colors_for_numeric(num_col, palette_type = "brewer",
  column_to_log = F, i = 1, granularity = 10)
}
\arguments{
\item{i}{}
}
\description{
generate a list of colors for the dataframe with all columns numeric (or date)
}
