% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures.R
\name{get_chain}
\alias{get_chain}
\title{Retrieve operation chain as vector or single string "combat_on_mediannorm_on_log"}
\usage{
get_chain(object, as_string = FALSE)
}
\arguments{
\item{object}{A `ProBatchFeatures` object.}

\item{as_string}{logical(1). if `TRUE` returns the chain as a single string
of the form `"combat_on_mediannorm_on_log"`.}
}
\value{
Character vector or string describing the processing chain.
}
\description{
Retrieve operation chain as vector or single string "combat_on_mediannorm_on_log"
}
