% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures copy.R, R/ProBatchFeatures.R
\name{get_operation_log}
\alias{get_operation_log}
\title{Access the operation log (structured)}
\usage{
get_operation_log(object)

get_operation_log(object)
}
\arguments{
\item{object}{A `ProBatchFeatures` object.}
}
\value{
S4Vectors::DataFrame

S4Vectors::DataFrame
}
\description{
Access the operation log (structured)

Access the operation log (structured)
}
