% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{get_peptide_corr_df}
\alias{get_peptide_corr_df}
\title{Transform square correlation matrix into long data frame of correlations}
\usage{
get_peptide_corr_df(peptide_cor, peptide_annotation,
  protein_col = "ProteinName", feature_id_col = "peptide_group_label")
}
\arguments{
\item{feature_id_col}{}
}
\description{
Transform square correlation matrix into long data frame of correlations
}
