% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{guess_factor_columns_if_needed}
\alias{guess_factor_columns_if_needed}
\title{Guess factors if numeric columns were not provided}
\usage{
guess_factor_columns_if_needed(
  factor_columns,
  sample_annotation,
  guess_factors
)
}
\arguments{
\item{factor_columns}{character vector of factor columns.}

\item{sample_annotation}{data frame of sample annotations.}

\item{guess_factors}{logical indicating whether to guess numeric columns.}
}
\value{
Named list containing updated factor_columns and numeric_columns.
}
\description{
Derive numeric columns from factor columns if guess_factors is TRUE
and numeric columns are NULL.
}
