% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{handle_factor_numeric_overlap}
\alias{handle_factor_numeric_overlap}
\title{Handle factor columns that are duplicated in numeric_columns}
\usage{
handle_factor_numeric_overlap(factor_columns, numeric_columns)
}
\arguments{
\item{factor_columns}{character vector of factor columns.}

\item{numeric_columns}{character vector of numeric columns.}
}
\value{
List with updated factor_columns and a warning if overlaps exist
}
\description{
Remove numeric columns from factor columns if overlap is detected.
}
