% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{join_data_matrices}
\alias{join_data_matrices}
\title{join list of matrices from different transformation steps into joined data frame}
\usage{
join_data_matrices(matrix_list, Step, sample_annotation,
  measure.col = "Intensity")
}
\arguments{
\item{measure.col}{}
}
\description{
join list of matrices from different transformation steps into joined data frame
}
