% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{log_transform}
\alias{log_transform}
\title{Log transformation of the data, ensuring that the row and column names
are retained}
\usage{
log_transform(data_matrix)
}
\arguments{
\item{data_matrix}{raw data matrix (features in rows and samples
in columns)}
}
\value{
`data_matrix`-size matrix, with columns log2 transformed
}
\description{
Log transformation of the data, ensuring that the row and column names
are retained
}
