% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{long_to_matrix}
\alias{long_to_matrix}
\title{Long to wide conversion}
\usage{
long_to_matrix(df_long, feature_id_col = "peptide_group_label",
  measure_col = "Intensity", sample_id_col = "FullRunName")
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}
}
\value{
\code{data_matrix} (\link{proBatch}) like matrix (features in rows, samples in columns)
}
\description{
Convert from a long data frame representation to a wide matrix representation
}
\seealso{
Other matrix manipulation functions: \code{\link{matrix_to_long}}
}
\concept{matrix manipulation functions}
