% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{matrix_to_long}
\alias{matrix_to_long}
\title{Wide to long conversion}
\usage{
matrix_to_long(data_matrix, sample_annotation = NULL,
  feature_id_col = "peptide_group_label", measure_col = "Intensity",
  sample_id_col = "FullRunName", step = NULL)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, 
use \code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with: 
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names) 
  \item biological covariates 
  \item technical covariates (batches etc) }. 
  See \code{help("example_sample_annotation")}}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, 
where the filenames (colnames of the \code{data_matrix} are found).}

\item{step}{normalization step (e.g. \code{Raw} or \code{Normalized}.
Useful if consecutive steps are compared in plots. Note
  that in plots these are usually ordered alphabetically, so it's worth
  naming with numbers, e.g. \code{1_raw}, \code{2_quantile}}
}
\value{
\code{df_long} (\link{proBatch}) like data frame
}
\description{
Convert from wide matrix to a long data frame representation
}
\examples{
proteome_long <- matrix_to_long(example_proteome_matrix, 
example_sample_annotation)

}
\seealso{
Other matrix manipulation functions: \code{\link{long_to_matrix}}
}
\concept{matrix manipulation functions}
