% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{matrix_to_long}
\alias{matrix_to_long}
\title{Convert the features x samples data matrix to a long format (e.g. for plotting)}
\usage{
matrix_to_long(data_matrix, sample_annotation, measure_col = "Intensity",
  step)
}
\arguments{
\item{sample_annotation}{}
}
\description{
Convert the features x samples data matrix to a long format (e.g. for plotting)
}
