% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{matrix_to_long}
\alias{matrix_to_long}
\title{Wide to long conversion}
\usage{
matrix_to_long(data_matrix, sample_annotation = NULL,
  feature_id_col = "peptide_group_label", measure_col = "Intensity",
  sample_id_col = "FullRunName", step = NULL)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{step}{normalization step (e.g. `Raw` or `Quantile_normalized` or
`qNorm_ComBat`). Useful if consecutive steps are compared in plots. Note
that in plots these are usually ordered alphabetically, so it's worth
naming with numbers, e.g. `1_raw`, `2_quantile`}
}
\value{
\code{df_long} (\link{proBatch}) like data frame
}
\description{
Convert from wide matrix to a long data frame representation
}
\seealso{
Other matrix manipulation functions: \code{\link{convert_to_matrix}},
  \code{\link{join_data_matrices}}
}
