% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{merge_rare_levels}
\alias{merge_rare_levels}
\title{Replaces rare levels with other}
\usage{
merge_rare_levels(column, rare_thr = 2)
}
\arguments{
\item{column}{column of the data whose rare categories need to be merged to
"other"}

\item{rare_thr}{minimal number of times for a category to be represented to
be declared as "rare" and converted to "other"}
}
\value{
column with rare occurrences replaced by other
}
\description{
Replaces levels with a maximal occurrence of 1 with \code{other}
}
\examples{
column <- factor(c("A", "B", "A", "C", "D", "D", "E"))
merge_rare_levels(column, rare_thr = 2)
# [1] A     other A     other D     D     other
# Levels: A D other
}
\keyword{internal}
