% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effect.R, R/normalize.R
\name{normalize}
\alias{normalize}
\alias{correct_medians_batch}
\alias{normalize_custom_fit}
\alias{correct_with_ComBat}
\alias{normalize_medians_global}
\title{Median normalization of the data (per batch median)}
\usage{
correct_medians_batch(df_long, sample_annotation = NULL,
  sample_id_col = "FullRunName", batch_col = "MS_batch",
  feature_id_col = "peptide_group_label", measure_col = "Intensity")

normalize_custom_fit(data_matrix, sample_annotation,
  batch_col = "MS_batch", feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName", measure_col = "Intensity",
  sample_order_col = "order", fit_func = fit_nonlinear, ...)

correct_with_ComBat(data_matrix, sample_annotation,
  sample_id_col = "FullRunName", batch_col = "MS_batch",
  par.prior = TRUE)

normalize_medians_global(df_long, sample_id_col = "FullRunName",
  measure_col = "Intensity")

normalize(data_matrix, normalizeFunc = "quantile", log = NULL)
}
\arguments{
\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be
repeated as row names) 2) biological and 3) technical covariates (batches
etc)}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{batch_col}{column in `sample_annotation` that should be used for
batch comparison}

\item{measure_col}{if `df_long` is among the parameters, it is the column
with expression/abundance/intensity, otherwise, it is used internally for
consistency}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{sample_order_col}{column, determining the order of sample MS run, used
as covariate to fit the non-linear fit}

\item{fit_func}{function to fit the (non)-linear trend}

\item{...}{other parameters, usually those of the `fit_func`}

\item{par.prior}{}

\item{normalizeFunc}{global batch normalization method (`quantile` or `MedianCentering`)}

\item{log}{whether to log transform data matrix before normalization (`NULL`, `2` or `10`)}

\item{return_long}{whether the result should be the "long" data frame (as
`df_long`) or "wide" (as `data_matrix`)}

\item{data_matrix}{raw data matrix (features in rows and samples
in columns)}
}
\value{
`data_matrix`-size data matrix with batch-effect corrected by
  `ComBat`

`data_matrix`-size matrix, with columns normalized
}
\description{
Median normalization of the data (per batch median)

normalize with the custom (continuous) fit

Standardized input-output ComBat normalization ComBat allows users to adjust
for batch effects in datasets where the batch covariate is known, using
methodology described in Johnson et al. 2007. It uses either parametric or
non-parametric empirical Bayes frameworks for adjusting data for batch
effects.  Users are returned an expression matrix that has been corrected for
batch effects. The input data are assumed to be cleaned and normalized before
batch effect removal.

Data normalization and batch adjustment methods

Median normalization of the data (global)

Batch correction method allows correction of continuous sigal drift within batch and 
discrete difference across batches.
}
\examples{

}
\seealso{
\code{\link{fit_nonlinear}}
}
