% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures copy.R, R/ProBatchFeatures.R
\name{pb_as_long}
\alias{pb_as_long}
\title{Get current assay as LONG (via proBatch::matrix_to_long)}
\usage{
pb_as_long(
  object,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  pbf_name = pb_current_assay(object)
)

pb_as_long(
  object,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  pbf_name = pb_current_assay(object)
)
}
\description{
Get current assay as LONG (via proBatch::matrix_to_long)

Get current assay as LONG (via proBatch::matrix_to_long)
}
