% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures.R
\name{pb_assay_matrix}
\alias{pb_assay_matrix}
\title{Convenience accessor for assay matrix by name/index (returns the 'intensity' assay)}
\usage{
pb_assay_matrix(object, assay = NULL, name = "intensity")
}
\description{
Convenience accessor for assay matrix by name/index (returns the 'intensity' assay)
}
