% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures copy.R, R/ProBatchFeatures.R
\name{pb_current_assay}
\alias{pb_current_assay}
\title{Current (latest) assay name}
\usage{
pb_current_assay(object)

pb_current_assay(object)
}
\description{
Current (latest) assay name

Current (latest) assay name
}
