% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures.R
\name{pb_eval}
\alias{pb_eval}
\title{Evaluate a pipeline and return the matrix, without storing}
\usage{
pb_eval(object, from, steps, funs = NULL, params_list = NULL)
}
\description{
Evaluate a pipeline and return the matrix, without storing
}
