% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures.R
\name{pb_has_step}
\alias{pb_has_step}
\title{Is a step available in the registry?}
\usage{
pb_has_step(name)
}
\arguments{
\item{name}{character(1) step name (e.g., "combat", "BERT")}
}
\value{
logical(1)
}
\description{
Is a step available in the registry?
}
\examples{
pb_has_step("medianNorm")
pb_has_step("BERT")
}
