% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_corr_between_samples}
\alias{plot_corr_between_samples}
\title{Sample correlation plot}
\usage{
plot_corr_between_samples(data_matrix, samples_to_plot, flavor = "corrplot",
  ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{samples_to_plot}{string vector of samples in \code{data_matrix} to be used in the plot}

\item{...}{parameters for the corrplot visualisation}
}
\description{
Plot correlation of selected samples
}
