% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_corr_matrix}
\alias{plot_corr_matrix}
\title{Visualise correlation matrix}
\usage{
plot_corr_matrix(corr_matrix, flavor = c("pheatmap", "corrplot"),
  filename = NULL, width = NA, height = NA, unit = c("cm", "in",
  "mm"), plot_title = NULL, ...)
}
\arguments{
\item{corr_matrix}{square correlation matrix}

\item{flavor}{either corrplot from 'corrplot' package or 
heatmap, as in 'pheatmap'}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only 
pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{unit}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{Title of the plot (usually, processing step + 
representation level (fragments, transitions, proteins))}

\item{...}{parameters for the \code{\link[corrplot]{corrplot.mixed}} or
\code{\link[pheatmap]{pheatmap}} visualisation, for details see examples and
  help to corresponding functions}
}
\value{
\code{corrplot} or \code{pheatmap} object depending on \code{flavor}
}
\description{
Plot correlation of selected  samples or peptides
}
\examples{
peptides <- c("10231_QDVDVWLWQQEGSSK_2", "10768_RLESELDGLR_2",
"1131_AQGSVALSVTQDPAR_2", "12111_TGKPNPDQLLK_2" , "1217_ASFQPPVC(UniMod:4)NK_2",                   "13503_VLEQGQHR_2"                              
"14845_YYVGDTEDVLFEK_2",  "15010_FANTM(UniMod:35)GLVIER_2", "27827_IM(UniMod:35)FVGGPNTR_2" ,
"30298_KDYHIEEGEEVFYQLEGDMILR_4" , "3347_ELPFPLNTR_2")

data_matrix_sub = example_proteome_matrix[peptides,]
corr_matrix = cor(t(data_matrix_sub), use = 'complete.obs')
plot_corr_matrix(corr_matrix,  flavor = "corrplot")
}
\seealso{
\code{\link[pheatmap]{pheatmap}}, \code{\link[corrplot]{corrplot.mixed}}
}
\keyword{internal}
