% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_corr_matrix}
\alias{plot_corr_matrix}
\title{Visualise correlation matrix}
\usage{
plot_corr_matrix(corr_matrix, annotation = NULL,
  annotation_id_col = "FullRunName", factors_to_plot = NULL,
  cluster_rows = FALSE, cluster_cols = FALSE,
  heatmap_color = colorRampPalette(rev(brewer.pal(n = 7, name =
  "RdYlBu")))(100), color_list = NULL, filename = NULL, width = 7,
  height = 7, units = c("cm", "in", "mm"), plot_title = NULL, ...)
}
\arguments{
\item{corr_matrix}{square correlation matrix}

\item{annotation}{data frame with \code{peptide_annotation} for protein 
correlation heatmap or \code{sample_annotation} for sample correlation heatmap}

\item{annotation_id_col}{\code{feature_id_col} for protein correlation heatmap 
or \code{sample_id_col} for sample correlation heatmap}

\item{factors_to_plot}{vector of technical and biological covariates to be 
plotted in this diagnostic plot (assumed to be present in 
\code{sample_annotation})}

\item{cluster_rows}{boolean values determining if rows should be clustered or \code{hclust} object}

\item{cluster_cols}{boolean values determining if columns should be clustered or \code{hclust} object}

\item{heatmap_color}{vector of colors used in heatmap.}

\item{color_list}{list, as returned by \code{sample_annotation_to_colors}, 
where each item contains a color vector for each factor to be mapped to the 
color.}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only 
pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{...}{parameters for the \code{\link[pheatmap]{pheatmap}} visualisation,
for details see examples and help to corresponding functions}
}
\value{
\code{pheatmap} object
}
\description{
recommended for heatmap-type visualisation of correlation matrix 
with <100 items. With >50 samples and ~10 replicate pairs distribution plots 
may be more informative.
}
\details{
Plot correlation of selected  samples or peptides
}
\examples{
peptides <- c("10231_QDVDVWLWQQEGSSK_2", "10768_RLESELDGLR_2")
data_matrix_sub = example_proteome_matrix[peptides,]
corr_matrix = cor(t(data_matrix_sub), use = 'complete.obs')
corr_matrix_plot <- plot_corr_matrix(corr_matrix)

}
\seealso{
\code{\link[pheatmap]{pheatmap}}, 
\code{\link{plot_sample_corr_distribution}}, 
\code{\link{plot_peptide_corr_distribution}}
}
