% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_corr_matrix}
\alias{plot_corr_matrix}
\title{Visualise correlation matrix}
\usage{
plot_corr_matrix(corr_matrix, flavor = "corrplot", filename = NULL,
  width = NA, height = NA, unit = c("cm", "in", "mm"),
  plot_title = "", ...)
}
\arguments{
\item{corr_matrix}{square correlation matrix}

\item{flavor}{either corrplot from 'corrplot' package or heatmap, as in 'pheatmap'}

\item{filename}{path where the results are saved. If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{unit}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{...}{parameters for the \code{\link[corrplot]{corrplot.mixed}} or
\code{\link[pheatmap]{pheatmap}} visualisation, for details see examples and
  help to corresponding functions}
}
\description{
Plot correlation of selected  samples or peptides
}
\seealso{
\code{\link[pheatmap]{pheatmap}}, \code{\link[corrplot]{corrplot.mixed}}
}
