% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_corr_plot_protein}
\alias{plot_corr_plot_protein}
\title{plot correlation plot of a single protein}
\usage{
plot_corr_plot_protein(data_matrix, protein_name, peptide_annotation,
  prot.column = "ProteinName", peptide_col_name = "peptide_group_label",
  title = NULL, ...)
}
\arguments{
\item{title}{}
}
\description{
plot correlation plot of a single protein
}
\examples{
plot_corr_plot(q_norm_proteome, protein_name = 'Haao',
peptide_annotation = peptide_annotation, prot.column = 'Gene',
title = 'Haao protein peptides after quantile norm',,
number.cex=0.75, tl.cex = .75
mar=c(0,0,1,0))
}
