% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of samples}
\usage{
plot_heatmap(data_matrix, sample_annotation, fill_the_missing = T,
  cluser_rows = F, cluster_cols = F, annotation_color_list = NA,
  filename = NA, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot the heatmap of samples
}
