% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap of samples}
\usage{
plot_heatmap(data_matrix, sample_annotation = NULL,
  fill_the_missing = T, cluster_rows = T, cluster_cols = F,
  annotation_color_list = NA,
  heatmap_color = colorRampPalette(rev(RColorBrewer::brewer.pal(n = 7,
  name = "RdYlBu")))(100), color_for_missing = "black", filename = NA,
  plot_title = NA, ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. in most
function, it is assumed that this is the log transformed version of the
original data}

\item{sample_annotation}{data matrix with \enumerate{
\item  `sample_id_col` (this can be repeated as row names)
  \item  biological and
  \item  technical covariates (batches etc)
}; each column of sample annotation will get it's own row. If `cluster_cols = T` this will indicate,
whether sample proximity is driven by one of biolical or technical factors}

\item{fill_the_missing}{boolean value determining if missing values should be
substituted with -1 (and colored with black)}

\item{cluster_rows}{boolean value determining if rows should be clustered}

\item{cluster_cols}{boolean value determining if columns should be clustered}

\item{annotation_color_list}{list specifying colors for columns (samples).
Best created by `sample_annotation_to_colors`}

\item{heatmap_color}{vector of colors used in heatmap (typicall a gradient)}

\item{color_for_missing}{special color to make missing values. Usually black or white, depending on `heatmap_color`}

\item{filename}{filepath where to save the image}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{...}{other parameters of \code{link[pheatmap]{pheatmap}}}
}
\value{
object returned by \code{link[pheatmap]{pheatmap}}
}
\description{
Plot the heatmap of samples
}
\seealso{
\code{\link{sample_annotation_to_colors}}, \code{\link[pheatmap]{pheatmap}}
}
