% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_hierarchical_clustering}
\alias{plot_hierarchical_clustering}
\title{cluster the data matrix to visually inspect which confounder dominates}
\usage{
plot_hierarchical_clustering(data_matrix, sample_annotation,
  sample_id_col = "FullRunName", color_list = NULL,
  factor_columns = c("MS_batch", "Diet"),
  numeric_columns = c("DateTime", "order"), fill_the_missing = 0,
  distance = "euclidean", agglomeration = "complete",
  label_samples = TRUE, label_font = 0.2, filename = NULL,
  width = 38, height = 25, units = c("cm", "in", "mm"),
  plot_title = NULL, ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{fill_the_missing}{boolean value determining if  missing values 
should be substituted with -1 (and colored with   \code{color_for_missing}). 
If \code{NULL}, features with missing values are excluded.}

\item{distance}{distance metric used for clustering}

\item{agglomeration}{agglomeration methods as used by \code{hclust}}

\item{label_samples}{if \code{TRUE} sample IDs (column names of 
\code{data_matrix}) will be printed}

\item{label_font}{size of the font. Is active if \code{label_samples} is 
\code{TRUE}, ignored otherwise}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only 
pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{...}{other parameters of \code{plotDendroAndColors} from \code{WGCNA} package}

\item{color_df}{data frame of colors, as created by 
\code{sample_annotation_to_colors}}
}
\value{
No return
}
\description{
cluster the data matrix to visually inspect which confounder dominates
}
\examples{
color_list <- sample_annotation_to_colors (example_sample_annotation, 
factor_columns = c('MS_batch','EarTag', "Strain", "Diet", "digestion_batch", "Sex"),
numeric_columns = c('DateTime', 'order'))

hierarchical_clustering_plot <- plot_hierarchical_clustering(
example_proteome_matrix, example_sample_annotation,
color_list = color_list,  
distance = "euclidean", agglomeration = 'complete',
label_samples = FALSE)

}
\seealso{
\code{\link[stats]{hclust}}, \code{\link{sample_annotation_to_colors}},
  \code{\link[WGCNA]{plotDendroAndColors}}
}
