% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_hierarchical_clustering}
\alias{plot_hierarchical_clustering}
\title{cluster the data matrix to visually inspect which confounder dominates}
\usage{
plot_hierarchical_clustering(data_matrix, color_df,
  fill_the_missing = NULL, distance = "euclidean",
  agglomeration = "complete", label_samples = TRUE, label_font = 0.2,
  filename = NULL, width = 38, height = 25, units = c("cm", "in",
  "mm"), plot_title = NULL, ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{color_df}{data frame of colors, as created by 
\code{sample_annotation_to_colors}}

\item{fill_the_missing}{boolean value determining if  missing values 
should be substituted with -1 (and colored with   \code{color_for_missing}). 
If \code{NULL}, features with missing values are excluded.}

\item{distance}{distance metric used for clustering}

\item{agglomeration}{agglomeration methods as used by \code{hclust}}

\item{label_samples}{if \code{TRUE} sample IDs (column names of 
\code{data_matrix}) will be printed}

\item{label_font}{size of the font. Is active if \code{label_samples} is 
\code{TRUE}, ignored otherwise}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only 
pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{...}{other parameters of \code{plotDendroAndColors} from \code{WGCNA} package}
}
\value{
No return
}
\description{
cluster the data matrix to visually inspect which confounder dominates
}
\examples{
color_scheme <- sample_annotation_to_colors (example_sample_annotation, 
factor_columns = c('MS_batch','EarTag', "Strain", "Diet", "digestion_batch", "Sex"),
date_columns = 'DateTime',
numeric_columns = c('order'))

color_annotation <- color_scheme$color_df

hiarchical_clustering_plot <- plot_hierarchical_clustering(
example_proteome_matrix, color_annotation,  
distance = "euclidean", agglomeration = 'complete',
label_samples = FALSE)

}
\seealso{
\code{\link[stats]{hclust}}, \code{\link{sample_annotation_to_colors}},
  \code{\link[WGCNA]{plotDendroAndColors}}
}
