% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing.R
\name{plot_NA_frequency}
\alias{plot_NA_frequency}
\alias{plot_NA_frequency.default}
\alias{plot_NA_frequency.ProBatchFeatures}
\title{Plot missing-value frequency distribution}
\usage{
plot_NA_frequency(x, ...)

\method{plot_NA_frequency}{default}(x, show_percent = FALSE, fill = "#345995", ...)

\method{plot_NA_frequency}{ProBatchFeatures}(
  x,
  pbf_name = NULL,
  fill = "#345995",
  nrow = NULL,
  ncol = NULL,
  facet_scales = "free_y",
  show_percent = FALSE,
  ...
)
}
\arguments{
\item{x}{A data container. For the `ProBatchFeatures` method this must be a
`ProBatchFeatures` object. The default method accepts any matrix-like input
(including `SummarizedExperiment`).}

\item{...}{Additional parameters forwarded to [pheatmap::pheatmap()].}

\item{fill}{Colour used for the columns in the frequency plot.}

\item{pbf_name}{Character scalar or vector with assay names to plot. When
`NULL`, the most recent assay returned by [pb_current_assay()] is used.
Only used by the `ProBatchFeatures` method.}

\item{nrow, ncol}{Integers controlling the layout when multiple assays are
plotted. If both are `NULL`, a roughly square layout is chosen
automatically.}

\item{facet_scales}{Scaling behaviour for facets when plotting multiple
assays.}
}
\value{
A `ggplot` object showing the frequency distribution.
}
\description{
Display how many features are observed in a given number of samples.
}
