% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{plot_pca}
\alias{plot_pca}
\title{plot PCA plot}
\usage{
plot_pca(data_matrix, sample_annotation, factor_to_color, PC_to_plot = c(1,
  2), colors_for_factor = NULL, theme = "classic")
}
\arguments{
\item{theme}{}
}
\description{
plot PCA plot
}
