% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{plot_pca}
\alias{plot_pca}
\title{plot PCA plot}
\usage{
plot_pca(data_matrix, sample_annotation, color_by = "MS_batch",
  PC_to_plot = c(1, 2), colors_for_factor = NULL, theme = "classic")
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. in most
function, it is assumed that this is the log transformed version of the
original data}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be
repeated as row names) 2) biological and 3) technical covariates (batches
etc)}

\item{color_by}{column name (as in `sample_annotation`) to color by}

\item{PC_to_plot}{principal component numbers for x and y axis}

\item{colors_for_factor}{named vector of colors for the `color_by` variable}

\item{theme}{ggplot theme, by default `classic`. Can be easily overriden (see
examples)}
}
\value{
ggplot scatterplot colored by factor levels of column specified in
  `factor_to_color`
}
\description{
plot PCA plot
}
