% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_pca}
\alias{plot_pca}
\title{plot PCA plot}
\usage{
plot_pca(data_matrix, sample_annotation, sample_id_col = "FullRunName",
  color_by = "MS_batch", PC_to_plot = c(1, 2), colors_for_factor = NULL,
  fill_the_missing = NULL, theme = "classic", plot_title = NULL)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. in most
function, it is assumed that this is the log transformed version of the
original data}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be
repeated as row names) 2) biological and 3) technical covariates (batches
etc)}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{color_by}{column name (as in `sample_annotation`) to color by}

\item{PC_to_plot}{principal component numbers for x and y axis}

\item{colors_for_factor}{named vector of colors for the `color_by` variable}

\item{fill_the_missing}{boolean value determining if missing values should be
substituted with -1 (and colored with black)}

\item{theme}{ggplot theme, by default `classic`. Can be easily overriden (see
examples)}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}
}
\value{
ggplot scatterplot colored by factor levels of column specified in
  `factor_to_color`
}
\description{
plot PCA plot
}
