% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_peptide_corr_distribution}
\alias{plot_peptide_corr_distribution}
\title{Plot distribution of peptide correlations within one 
protein and between proteins}
\usage{
plot_peptide_corr_distribution(data_matrix, peptide_annotation,
  protein_col = "ProteinName", feature_id_col = "peptide_group_label",
  plot_title = "Distribution of peptide correlation",
  theme = "classic")
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{peptide_annotation}{long format data frame with peptide ID and their corresponding 
protein and/or gene annotations. 
See "example_peptide_annotation" for more details (to call the description, use \code{help("example_peptide_annotation")})}

\item{protein_col}{column where protein names are specified}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{plot_title}{title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}
}
\value{
\code{ggplot} type object with violin plot
}
\description{
Plot distribution of peptide correlations within one 
protein and between proteins
}
\examples{
peptide_corr_distribution <- plot_peptide_corr_distribution(
example_proteome_matrix, 
example_peptide_annotation, protein_col = 'Gene')

}
