% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_peptide_corr_distribution}
\alias{plot_peptide_corr_distribution}
\title{Plot distribution of peptide correlations within one 
protein and between proteins}
\usage{
plot_peptide_corr_distribution(data_matrix, peptide_annotation,
  protein_col = "ProteinName", feature_id_col = "peptide_group_label",
  plot_title = "Distribution of peptide correlation",
  theme = "classic")
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{peptide_annotation}{long format data with peptide ID
and their corresponding 
protein annotations}

\item{protein_col}{the column name in \code{peptide_annotation} 
with protein names}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{plot_title}{Title of the plot, usually processing step}

\item{theme}{ggplot theme, by default `classic`. Can be easily overriden}

\item{\code{...}}{parameters for the \code{ggplot} visualisation}
}
\value{
\code{ggplot} type object with violin 
plot for each \code{plot_param}
}
\description{
Plot distribution of peptide correlations within one 
protein and between proteins
}
\examples{
plot_peptide_corr_distribution(example_proteome_matrix, 
example_peptide_annotation, protein_col = 'Gene')

}
