% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_peptide_level}
\alias{plot_peptide_level}
\title{Plot peptide measurements}
\usage{
plot_peptide_level(pep_name, df_long, sample_annotation, order_column = NULL,
  sample_id_col = "FullRunName", batch_column = "MS_batch.final",
  measure_column = "Intensity", feature_id_column = "peptide_group_label",
  geom = c("point", "line"), color_by_batch = F, facet_by_batch = F,
  title = NULL, requant = NULL, theme = "classic")
}
\arguments{
\item{pep_name}{name of the peptide for diagnostic profiling}

\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_column} and a
\code{measure_column}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{order_column}{column in \code{sample_annotation} that determines sample order. It is
used for certain diagnostics and normalisations.}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{batch_column}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{measure_column}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{feature_id_column}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{geom}{whether to show the feature as points and/or connect by lines}

\item{color_by_batch}{(logical) whether to color points by batch}

\item{facet_by_batch}{(logical) whether to plot each batch in its own facet}

\item{title}{the string indicating the source of the peptides}

\item{requant}{if data frame: requant values; if logical: whether to indicate
requant values (requires 'requant' or 'm_score' column in \code{df_long})}

\item{theme}{plot theme (default is 'classical'; other options not
implemented)}
}
\value{
ggplot2 type plot of \code{measure_column} vs \code{order_column},
  faceted by \code{pep_name} and (optionally) by \code{batch_column}
}
\description{
Creates a peptide facetted ggplot2 plot of the value in \code{measure_column}
vs \code{order_column}. Additionally, the resulting plot can also be facetted
by batch.
}
\seealso{
Other feature-level diagnostic functions: \code{\link{plot_iRTs}},
  \code{\link{plot_spike_ins}},
  \code{\link{plot_with_fitting_curve}}
}
