% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_peptides_of_one_protein}
\alias{plot_peptides_of_one_protein}
\title{Plot peptides of one protein}
\usage{
plot_peptides_of_one_protein(protein_name, peptide_annotation = NULL,
  protein_col = "ProteinName", df_long, sample_annotation = NULL,
  sample_id_col = "FullRunName", measure_col = "Intensity",
  feature_id_col = "peptide_group_label", geom = c("point", "line"),
  qual_col = NULL, qual_value = NULL, batch_col = "MS_batch",
  color_by_batch = FALSE, color_scheme = "brewer",
  order_col = "order", vline_color = "red", facet_col = NULL,
  plot_title = sprintf("Peptides of \%s protein", protein_name),
  theme = "classic", ...)
}
\arguments{
\item{protein_name}{name of the protein as defined in \code{ProteinName}}

\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file). See "example_proteome" for more details (to call the description, use \code{help("example_proteome")})}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{geom}{whether to show the feature as points and/or connect by lines (accepted 
values are: 1. \code{point}, \code{line} and \code{c('point', 'line')})}

\item{qual_col}{column to color point by certain value denoted 
by \code{color_by_qual_value}. Design with inferred/requant values in OpenSWATH output data, 
which means argument value has to be set to \code{m_score}.}

\item{qual_value}{value in \code{qual_col} to color. For OpenSWATH data,
this argument value has to be set to \code{2} (this is an \code{m_score} value for requants).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{color_by_batch}{(logical) whether to color points and connecting lines 
by batch factor as defined by \code{batch_col}.}

\item{order_col}{column in \code{sample_annotation} that determines sample order. It is
used for in initial assessment plots (\link{plot_sample_mean_or_boxplot}) and 
feature-level diagnostics (\link{feature_level_diagnostics}). Can be `NULL` 
if sample order is irrelevant (e.g. in genomic experiments). For more details,
order definition/inference, see \link{define_sample_order} and \link{date_to_sample_order}}

\item{vline_color}{color of vertical lines, typically separating 
different MS batches in ordered runs; 
should be `NULL` for experiments without intrinsic order.}

\item{facet_col}{column  in \code{sample_annotation} with a batch factor to separate 
plots into facets; usually 2nd to \code{batch_col}. Most meaningful for multi-instrument 
MS experiments (where each instrument has its own order-associated effects, see \code{order_col}) 
or simultaneous examination of two batch factors (e.g. preparation day and measurement day).
For single-instrument case should be set to `NULL`}

\item{plot_title}{title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}

\item{...}{additional arguments to \code{\link{plot_single_feature}} function}
}
\value{
ggplot2 type plot of \code{measure_col} vs \code{order_col},
  faceted by \code{spike_ins} containing proteins and (optionally) by \code{batch_col}
}
\description{
Creates a spike-in faceted ggplot2 plot of the value in
\code{measure_col} vs \code{order_col} using
\code{\link{plot_single_feature}}. Additionally, the resulting plot can also
be faceted by batch.
}
\examples{
peptides_of_one_protein_plot <- plot_peptides_of_one_protein (
protein_name = "Haao",  
protein_col = "Gene", df_long = example_proteome, 
example_sample_annotation, 
order_col = 'order', sample_id_col = 'FullRunName', 
batch_col = 'MS_batch')

}
\seealso{
Other feature-level diagnostic functions: \code{\link{feature_level_diagnostics}},
  \code{\link{plot_iRT}}, \code{\link{plot_spike_in}},
  \code{\link{plot_with_fitting_curve}}
}
\concept{feature-level diagnostic functions}
