% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_peptides_of_one_protein}
\alias{plot_peptides_of_one_protein}
\title{Plot peptides of one protein}
\usage{
plot_peptides_of_one_protein(protein_name, protein_col = "ProteinName",
  df_long, sample_annotation, peptide_annotation = NULL,
  order_col = "order", sample_id_col = "FullRunName",
  batch_col = "MS_batch", measure_col = "Intensity",
  feature_id_col = "peptide_group_label", color_by_col = NULL,
  color_by_value = NULL,
  plot_title = sprintf("Peptides of \%s protein", protein_name), ...)
}
\arguments{
\item{protein_name}{name of the protein as defined in \code{ProteinName}}

\item{protein_col}{column where protein names are specified}

\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{order_col}{column in \code{sample_annotation} that determines sample order. It is
used for certain diagnostics and normalisations.}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{plot_title}{the string indicating the source of the peptides}

\item{...}{additional arguments to \code{\link{plot_single_feature}} function}
}
\value{
ggplot2 type plot of \code{measure_col} vs \code{order_col},
  faceted by \code{spike_ins} containing proteins and (optionally) by \code{batch_col}
}
\description{
Creates a spike-in facetted ggplot2 plot of the value in
\code{measure_col} vs \code{order_col} using
\code{\link{plot_single_feature}}. Additionally, the resulting plot can also
be facetted by batch.
}
\seealso{
Other feature-level diagnostic functions: \code{\link{plot_iRT}},
  \code{\link{plot_single_feature}},
  \code{\link{plot_spike_in}},
  \code{\link{plot_with_fitting_curve}}
}
\concept{feature-level diagnostic functions}
