% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_protein_corrplot}
\alias{plot_protein_corrplot}
\title{Peptide correlation matrix (heatmap)}
\usage{
plot_protein_corrplot(data_matrix, protein_name,
  peptide_annotation = NULL, protein_col = "ProteinName",
  feature_id_col = "peptide_group_label", flavor = c("pheatmap",
  "corrplot"), filename = NULL, width = NA, height = NA,
  units = c("cm", "in", "mm"),
  plot_title = sprintf("Peptide correlation matrix of \%s protein",
  protein_name), ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{protein_name}{the name of the protein}

\item{peptide_annotation}{long format data frame with peptide ID and their corresponding 
protein and/or gene annotations. 
See "example_peptide_annotation" for more details (to call the description, use \code{help("example_peptide_annotation")})}

\item{protein_col}{column where protein names are specified}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{flavor}{either corrplot from 'corrplot' 
package or heatmap, as in 'pheatmap'}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently 
only pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{The title of the plot, e.g. protein name / processing step}

\item{...}{parameters for the corrplot visualisation}
}
\value{
\code{corrplot} or \code{pheatmap} object depending on \code{flavor}
}
\description{
Plots correlation plot of peptides from a single protein
}
\examples{
protein_corrplot_plot <- plot_protein_corrplot(example_proteome_matrix, 
protein_name = 'Haao', peptide_annotation = example_peptide_annotation, 
protein_col = 'Gene', flavor = "pheatmap")

}
