% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_protein_corrplot}
\alias{plot_protein_corrplot}
\title{Peptide correlation matrix (heatmap)}
\usage{
plot_protein_corrplot(data_matrix, protein_name, peptide_annotation,
  protein_col = "ProteinName", feature_id_col = "peptide_group_label",
  flavor = "corrplot", filename = NULL, width = NA, height = NA,
  unit = c("cm", "in", "mm"),
  plot_title = sprintf("Peptide correlation matrix of \%s protein",
  protein_name), ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{protein_name}{the name of the protein}

\item{peptide_annotation}{df with peptides and their corresponding proteins}

\item{protein_col}{the column name in \code{peptide_annotation} with protein names}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{flavor}{either corrplot from 'corrplot' package or heatmap, as in 'pheatmap'}

\item{filename}{path where the results are saved. If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{unit}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{The title of the plot}

\item{...}{parameters for the corrplot visualisation}

\item{peptide_col_name}{the column name in \code{peptide_annotation} with peptide names}
}
\description{
Plots correlation plot of peptides from a single protein
}
\examples{
\\donotrun{plot_corr_plot(q_norm_proteome, protein_name = 'Haao',
               peptide_annotation = peptide_annotation, prot.column = 'Gene',
               title = 'Haao protein peptides after quantile norm',
               number.cex=0.75, tl.cex = .75
               mar=c(0,0,1,0))}
\\donotrun{lower = "ellipse", upper = "number",
 tl.col = "black", diag = 'l', tl.pos = "lt", number.cex=0.75, tl.cex = .75}

}
