% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{plot_pvca}
\alias{plot_pvca}
\title{Quantify variance distribution by variable}
\usage{
plot_pvca(data_matrix, sample_annotation, sample_id_column = "FullRunName",
  technical_covariates = NULL, biological_covariates = NULL,
  plot_title = NULL, colors_for_bars = NULL, threshold_pca = 0.6,
  threshold_var = 0.01, theme = "classic")
}
\arguments{
\item{threshold_var}{}
}
\description{
Quantify variance distribution by variable
}
