% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{plot_pvca}
\alias{plot_pvca}
\title{Plot variance distribution by variable}
\usage{
plot_pvca(data_matrix, sample_annotation, sample_id_col = "FullRunName",
  technical_covariates = c("MS_batch", "instrument"),
  biological_covariates = c("cell_line", "drug_dose"), title = NULL,
  colors_for_bars = NULL, threshold_pca = 0.6, threshold_var = 0.01,
  theme = "classic")
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. in most
function, it is assumed that this is the log transformed version of the
original data}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be
repeated as row names) 2) biological and 3) technical covariates (batches
etc)}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{technical_covariates}{vector `sample_annotation` column names that are
technical covariates}

\item{biological_covariates}{vector `sample_annotation` column names, that
are biologically meaningful covariates}

\item{title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{colors_for_bars}{four-item color vector, specifying colors for the
following categories: c('residual', 'biological', 'biol:techn',
'technical')}

\item{threshold_pca}{the percentile value of the minimum amount of the
variabilities that the selected principal components need to explain}

\item{threshold_var}{the percentile value of weight each of the covariates
needs to explain (the rest will be lumped together)}

\item{theme}{ggplot theme, by default `classic`. Can be easily overriden (see
examples)}
}
\value{
list of two items: plot =gg, df = pvca_res
}
\description{
Plot variance distribution by variable
}
\seealso{
\code{\link{sample_annotation_to_colors}}
}
