% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_sample_corr_distribution}
\alias{plot_sample_corr_distribution}
\alias{plot_sample_corr_distribution.corrDF}
\title{Create violin plot of sample correlation distribution}
\usage{
plot_sample_corr_distribution(data_matrix, sample_annotation,
  repeated_samples = NULL, sample_id_col = "FullRunName",
  batch_col = "MS_batch", biospecimen_id_col = "EarTag",
  plot_title = "Sample correlation distribution",
  plot_param = "batch_replicate", theme = "classic")

plot_sample_corr_distribution.corrDF(corr_distribution,
  plot_title = "Sample correlation distribution",
  plot_param = "batch_replicate", theme = "classic")
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{repeated_samples}{if \code{NULL}, correlation of all samples is plotted}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{biospecimen_id_col}{column in \code{sample_annotation} 
that captures the biological sample, 
that (possibly) was profiled several times as technical replicates.
Tip: if such ID is absent, but can be defined from several columns,
create new \code{biospecimen_id} column}

\item{plot_title}{title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{plot_param}{columns, defined in correlation_df, which is output of
\code{calculate_sample_corr_distribution}, specifically,  \enumerate{
\item \code{replicate}
\item \code{batch_the_same}
\item \code{batch_replicate}
\item \code{batches}
}}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}

\item{corr_distribution}{data frame with correlation distribution, 
as returned by \code{calculate_sample_corr_distribution}}
}
\value{
\code{ggplot} type object with violin plot 
for each \code{plot_param}
}
\description{
Useful to visualize within batch vs within replicate 
vs non-related sample correlation
}
\examples{
sample_corr_distribution_plot <- plot_sample_corr_distribution(
example_proteome_matrix,
example_sample_annotation, batch_col = 'MS_batch', 
biospecimen_id_col = "EarTag", 
plot_param = 'batch_replicate')

corr_distribution = calculate_sample_corr_distribution(data_matrix = example_proteome_matrix, 
sample_annotation = example_sample_annotation,
batch_col = 'MS_batch',biospecimen_id_col = "EarTag")
sample_corr_distribution_plot <- plot_sample_corr_distribution.corrDF(corr_distribution,
plot_param = 'batch_replicate')

}
\seealso{
\code{\link{calculate_sample_corr_distribution}}, \code{\link[ggplot2]{ggplot}}
}
