% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_sample_corr_distribution}
\alias{plot_sample_corr_distribution}
\title{Create violin plot of correlation distribution}
\usage{
plot_sample_corr_distribution(data_matrix, sample_annotation,
  repeated_samples = NULL, sample_id_col = "FullRunName",
  batch_col = "MS_batch", biospecimen_id_col = "EarTag",
  plot_title = "Correlation distribution",
  plot_param = "batch_replicate")
}
\arguments{
\item{repeated_samples}{if `NULL`, only repeated sample correlation is plotted}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix) are found}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{biospecimen_id_col}{column in `sample_annotation` that captures the biological sample, 
that (possibly) was profiled several times as technical replicates.
Tip: if such ID is absent, but can be defined from several columns,
create new \code{biospecimen_id} column}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{plot_param}{columns, defined in correlation_df, which is output of
\code{get_sample_corr_distrib}, specifically, #' \enumerate{
\item \code{replicate}
\item \code{batch_the_same}
\item \code{batch_replicate}
\item \code{batches}
};}
}
\value{
\code{ggplot} type object with violin plot for each \code{plot_param}
}
\description{
Useful to visualize within batch vs within replicate vs non-related sample correlation
}
\examples{
\dontrun{plot_sample_corr_distribution(quantile_normalized_matrix,
example_sample_annotation, batch_col = 'MS_batch', biospecimen_id_col = "EarTag", 
plot_title = 'Quantile normalized',plot_param = 'batch_replicate')}
}
\seealso{
\code{\link{get_sample_corr_distrib}}, \code{\link[ggplot2]{ggplot}}
}
