% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_sample_corr_distribution}
\alias{plot_sample_corr_distribution}
\title{Create violin plot of correlation distribution
Typically to visualize within batch vs within replicate vs non-related sample correlation}
\usage{
plot_sample_corr_distribution(data_matrix, sample_annotation,
  repeated_samples = NULL, sample_id_col = "FullRunName",
  batch_col = "MS_batch.final", covariate = "EarTag",
  title = "Correlation_distribution", plot_param = "batch_the_same")
}
\arguments{
\item{plot_param}{}
}
\description{
Create violin plot of correlation distribution
Typically to visualize within batch vs within replicate vs non-related sample correlation
}
\examples{

}
