% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_sample_corr_heatmap}
\alias{plot_sample_corr_heatmap}
\title{Sample correlation matrix (heatmap)}
\usage{
plot_sample_corr_heatmap(data_matrix, samples_to_plot = NULL,
  sample_annotation = NULL, sample_id_col = "FullRunName",
  flavor = c("pheatmap", "corrplot"), filename = NULL, width = NA,
  height = NA, units = c("cm", "in", "mm"),
  plot_title = sprintf("Correlation matrix of\%s samples",
  ifelse(is.null(samples_to_plot), "", " selected")), ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, use \code{help("example_proteome_matrix")})}

\item{samples_to_plot}{string vector of samples in 
\code{data_matrix} to be used in the plot}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{flavor}{either corrplot from 'corrplot' package or 
heatmap, as in 'pheatmap'}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. Currently only 
pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{units}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{...}{parameters for the \code{\link[corrplot]{corrplot.mixed}} or
\code{\link[pheatmap]{pheatmap}} visualisation, for details see 
  examples and help to corresponding functions}
}
\value{
\code{corrplot} or \code{pheatmap} object depending on \code{flavor}
}
\description{
Plot correlation of selected samples
}
\examples{
specified_samples = example_sample_annotation$FullRunName[
which(example_sample_annotation$order \%in\% 110:115)] 

sample_corr_heatmap <- plot_sample_corr_heatmap(example_proteome_matrix, 
samples_to_plot = specified_samples, 
 flavor = 'pheatmap', 
 cluster_rows= FALSE, cluster_cols=FALSE,
 annotation_names_col = TRUE, annotation_legend = FALSE, 
 show_colnames = FALSE)

}
\seealso{
\code{\link[pheatmap]{pheatmap}}, 
\code{\link[corrplot]{corrplot.mixed}}
}
