% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-based_diagnostics.R
\name{plot_sample_corr_heatmap}
\alias{plot_sample_corr_heatmap}
\title{Sample correlation matrix (heatmap)}
\usage{
plot_sample_corr_heatmap(data_matrix, samples_to_plot = NULL,
  flavor = "corrplot", filename = NULL, width = NA, height = NA,
  unit = c("cm", "in", "mm"),
  plot_title = sprintf("Correlation matrix of sample \%s",
  samples_to_plot), ...)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{samples_to_plot}{string vector of samples in 
\code{data_matrix} to be used in the plot}

\item{flavor}{either corrplot from 'corrplot' package or 
heatmap, as in 'pheatmap'}

\item{filename}{path where the results are saved. 
If null the object is returned to the active window;
otherwise, the object is save into the file. 
Currently only pdf and png format is supported}

\item{width}{option  determining the output image width}

\item{height}{option  determining the output image width}

\item{unit}{units: 'cm', 'in' or 'mm'}

\item{plot_title}{Title of the plot (usually, processing step + 
representation level (fragments, transitions, proteins))}

\item{...}{parameters for the \code{\link[corrplot]{corrplot.mixed}} or
\code{\link[pheatmap]{pheatmap}} visualisation, for details see 
  examples and help to corresponding functions}
}
\value{
Nothing
}
\description{
Plot correlation of selected samples
}
\examples{
specified_samples = sample_annotation \%>\%
filter(RunID \%in\% paste('Run', 110:115, sep = '')) \%>\%
pull(FullRunName)

plot_sample_corr_heatmap(example_proteome_matrix, 
samples_to_plot = specified_samples, 
 flavor = 'pheatmap', 
 cluster_rows= FALSE, cluster_cols=FALSE,fontsize = 4,
 annotation_names_col = TRUE, annotation_legend = FALSE, 
 show_colnames = FALSE)

plot_sample_corr_heatmap(data_matrix, 
sample_to_plot = specified_samples,
 flavor = 'corrplot', lower = "ellipse", upper = "number")

}
\seealso{
\code{\link[pheatmap]{pheatmap}}, 
\code{\link[corrplot]{corrplot.mixed}}
}
