% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_diagnostics.R
\name{plot_sample_mean}
\alias{plot_sample_mean}
\title{Plot the sample average}
\usage{
plot_sample_mean(data_matrix, sample_annotation,
  sample_id_col = "FullRunName", order_column = "order",
  batch_column = NULL, color_by_batch = F, theme = "classic",
  title = NULL, color_scheme = "brewer")
}
\arguments{
\item{color_scheme}{}
}
\description{
Plot the sample average
}
