% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_spike_ins}
\alias{plot_spike_ins}
\title{Plot spike-in measurements}
\usage{
plot_spike_ins(df_long, sample_annotation, order_column = "order",
  spike_ins = "BOVIN", sample_id_column = "FullRunName",
  batch_column = "MS_batch", measure_column = "Intensity",
  feature_id_column = "peptide_group_label",
  title = "Spike-in BOVINE protein peptides", ...)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_column} and a
\code{measure_column}, but usually also an \code{m_score} (in OpenSWATH output result
file)}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{order_column}{column in \code{sample_annotation} that determines sample order. It is
used for certain diagnostics and normalisations.}

\item{spike_ins}{substring used to identify spike-in proteins in the column
'ProteinName'}

\item{batch_column}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{measure_column}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{feature_id_column}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{title}{the string indicating the source of the peptides}

\item{...}{additional arguments to \code{\link{plot_peptide_level}} function}
}
\value{
ggplot2 type plot of \code{measure_column} vs \code{order_column},
  faceted by \code{spike_ins} containing proteins and (optionally) by \code{batch_column}
}
\description{
Creates a spike-in facetted ggplot2 plot of the value in
\code{measure_column} vs \code{order_column} using
\code{\link{plot_peptide_level}}. Additionally, the resulting plot can also
be facetted by batch.
}
\seealso{
Other feature-level diagnostic functions: \code{\link{plot_iRTs}},
  \code{\link{plot_peptide_level}},
  \code{\link{plot_with_fitting_curve}}
}
