% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split_violinplot.R
\name{plot_split_violin_with_boxplot}
\alias{plot_split_violin_with_boxplot}
\title{Plot split violin plot (convenient to compare distribution before and after)}
\usage{
plot_split_violin_with_boxplot(df, y_col = "y", col_for_color = "m",
  col_for_box = "x", colors_for_plot = c("#8f1811", "#F8C333"),
  hlineintercept = NULL, plot_title = NULL, theme = "classic")
}
\arguments{
\item{df}{data.frame with \code{y_col}, \code{col_for_color}, \code{col_for_box}}

\item{y_col}{value to explore the distribution of}

\item{col_for_color}{column to use to map to two colors}

\item{col_for_box}{column to use to do group comparison}

\item{colors_for_plot}{colors to map to col_for_color}

\item{hlineintercept}{NULL: no intercept line; non-null: intercept value}

\item{plot_title}{title of the plot (e.g., processing step + representation
level (fragments, transitions, proteins) + purpose (meanplot/corrplot etc))}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}
}
\value{
ggplot object
}
\description{
Plot split violin plot (convenient to compare distribution before and after)
}
