% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_with_fitting_curve}
\alias{plot_with_fitting_curve}
\title{Plot peptide measurements across multi-step analysis}
\usage{
plot_with_fitting_curve(feature_name, fit_df, fit_value_var = "fit",
  df_long, sample_annotation = NULL, sample_id_col = "FullRunName",
  measure_col = "Intensity", feature_id_col = "peptide_group_label",
  geom = c("point", "line"), qual_col = NULL, qual_value = NULL,
  batch_col = "MS_batch", color_by_batch = FALSE,
  color_scheme = "brewer", order_col = "order", vline_color = "grey",
  facet_col = NULL,
  plot_title = sprintf("Fitting curve of \%s peptide", feature_name),
  theme = "classic")
}
\arguments{
\item{feature_name}{name of the selected feature (e.g. peptide) for diagnostic profiling}

\item{fit_df}{data frame typically output generated from nonlinear curve 
fitting by \code{normalize_custom_fit}}

\item{fit_value_var}{column denoting intensity values, typically fitted to curve}

\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} and a
\code{measure_col}, but usually also an \code{m_score} (in OpenSWATH output result
file). See "example_proteome" for more details (to call the description, use \code{help("example_proteome")})}

\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{geom}{whether to show the feature as points and/or connect by lines (accepted 
values are: 1. \code{point}, \code{line} and \code{c('point', 'line')})}

\item{qual_col}{column to color point by certain value denoted 
by \code{color_by_qual_value}. Design with inferred/requant values in OpenSWATH output data, 
which means argument value has to be set to \code{m_score}.}

\item{qual_value}{value in \code{qual_col} to color. For OpenSWATH data,
this argument value has to be set to \code{2} (this is an \code{m_score} value for imputed values (requant values).}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{color_by_batch}{(logical) whether to color points and connecting lines 
by batch factor as defined by \code{batch_col}.}

\item{order_col}{column in \code{sample_annotation} that determines sample order. It is
used for in initial assessment plots (\link{plot_sample_mean_or_boxplot}) and 
feature-level diagnostics (\link{feature_level_diagnostics}). Can be `NULL` 
if sample order is irrelevant (e.g. in genomic experiments). For more details,
order definition/inference, see \link{define_sample_order} and \link{date_to_sample_order}}

\item{vline_color}{color of vertical lines, typically separating 
different MS batches in ordered runs; 
should be `NULL` for experiments without intrinsic order.}

\item{facet_col}{column  in \code{sample_annotation} with a batch factor to separate 
plots into facets; usually 2nd to \code{batch_col}. Most meaningful for multi-instrument 
MS experiments (where each instrument has its own order-associated effects, see \code{order_col}) 
or simultaneous examination of two batch factors (e.g. preparation day and measurement day).
For single-instrument case should be set to `NULL`}

\item{plot_title}{title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{theme}{ggplot theme, by default \code{classic}. Can be easily overriden}

\item{...}{additional arguments to \code{\link{plot_single_feature}} function}
}
\value{
\code{ggplot}-class plot with minimally two facets (before and after
  non-linear fit) with \code{measure_col} (Intensity) vs \code{order_col}
  (injection order) for selected peptides (specified in \code{feature_name})
}
\description{
Plot Intensity of a few representative peptides for each step of the analysis
including the fitting curve
}
\examples{
loess_fit_70 <- adjust_batch_trend(example_proteome, 
example_sample_annotation, span = 0.7)

fitting_curve_plot <- plot_with_fitting_curve(feature_name = "10231_QDVDVWLWQQEGSSK_2", 
df_long = example_proteome, example_sample_annotation, 
fit_df = loess_fit_70$fit_df, plot_title = "Curve fitting with 70\% span")

}
\seealso{
Other feature-level diagnostic functions: \code{\link{feature_level_diagnostics}},
  \code{\link{plot_iRT}},
  \code{\link{plot_peptides_of_one_protein}},
  \code{\link{plot_spike_in}}
}
\concept{feature-level diagnostic functions}
