% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_with_fitting_curve}
\alias{plot_with_fitting_curve}
\title{Plot Intensity for a few representative peptides for each step of the analysis including the fitting curve}
\usage{
plot_with_fitting_curve(pep_name, data_df_all_steps, fit_df, sample_annotation,
  fit_value_var = "fit", fit_step = "3_loess_fit",
  batch_column = "MS_batch", feature_id_column = "peptide_group_label",
  measurement.col = "Intensity", sample_id_column = "FullRunName",
  order_column = NULL, geom = c("point", "line"), color_by_batch = F,
  facet_by_batch = F, title = NULL, requant = NULL, theme = "classic",
  color_var = "fit")
}
\arguments{
\item{color_var}{}
}
\description{
Plot Intensity for a few representative peptides for each step of the analysis including the fitting curve
}
