% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_level_diagnostics.R
\name{plot_with_fitting_curve}
\alias{plot_with_fitting_curve}
\title{Plot peptide measurements across multi-step analysis}
\usage{
plot_with_fitting_curve(pep_name, data_df_all_steps, sample_annotation,
  fit_df, fit_value_var = "fit", fit_step = "3_loess_fit",
  order_col = "order", sample_id_col = "FullRunName",
  batch_col = "MS_batch", measure_col = "Intensity",
  feature_id_col = "peptide_group_label", geom = c("point", "line"),
  color_by_batch = F, facet_by_batch = F,
  plot_title = sprintf("Fitting curve of \%s peptide", pep_name),
  color_by_col = NULL, color_by_value = NULL, theme = "classic",
  vline_color = "grey", ...)
}
\arguments{
\item{pep_name}{name of the peptide for diagnostic profiling}

\item{data_df_all_steps}{data frame, similar to \code{df_long}
\link{proBatch},  where each row is a single feature in a single sample, at
a certain step of the analysis (minimally raw and after linear
normalization) thus it has minimally the following columns:
\code{sample_id_col}, \code{feature_id_col}, \code{measure_col}, and
\code{fit_step}, but usually also \code{m_score}}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{order_col}{column in \code{sample_annotation} that determines sample order. It is
used for certain diagnostics and normalisations.}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{batch_col}{column in \code{sample_annotation} that should be used for
batch comparison}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{geom}{for the intensity \code{measure_col} profile:}

\item{color_by_batch}{(logical) whether to color points by batch}

\item{facet_by_batch}{(logical) whether to plot each batch in its own facet}

\item{plot_title}{the string indicating the source of the peptides}

\item{theme}{plot theme (default is 'classical'; other options not
implemented)}
}
\value{
\code{ggplot}-class plot with minimally two facets (before and after
  non-linear fit) with \code{measure_col} (Intensity) vs \code{order_col}
  (injection order) for selected peptides (specified in \code{pep_name})
}
\description{
Plot Intensity of a few representative peptides for each step of the analysis
including the fitting curve
}
\seealso{
Other feature-level diagnostic functions: \code{\link{plot_iRT_trend}},
  \code{\link{plot_peptide_trend}},
  \code{\link{plot_peptides_of_one_protein}},
  \code{\link{plot_spike_in_protein}}
}
\concept{feature-level diagnostic functions}
