% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_wide_diagnostics.R
\name{prepare_PVCA_df}
\alias{prepare_PVCA_df}
\title{prepare the weights of Principal Variance Components}
\usage{
prepare_PVCA_df(data_matrix, sample_annotation,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName", technical_factors = c("MS_batch",
  "instrument"), biological_factors = c("cell_line", "drug_dose"),
  fill_the_missing = -1, pca_threshold = 0.6,
  variance_threshold = 0.01)
}
\arguments{
\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, 
use \code{help("example_proteome_matrix")})}

\item{sample_annotation}{data frame with: 
\enumerate{ \item \code{sample_id_col} (this can be repeated as row names) 
  \item biological covariates 
  \item technical covariates (batches etc) }. 
  See \code{help("example_sample_annotation")}}

\item{feature_id_col}{name of the column with feature/gene/peptide/protein
ID used in the long format representation \code{df_long}. In the wide
formatted representation \code{data_matrix} this corresponds to the row
names.}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, 
where the filenames (colnames of the \code{data_matrix} are found).}

\item{technical_factors}{vector \code{sample_annotation} column names that 
are technical covariates}

\item{biological_factors}{vector \code{sample_annotation} column names, that
are biologically meaningful covariates}

\item{fill_the_missing}{numeric value determining how  missing values 
should be substituted. If \code{NULL}, features with missing values are 
excluded.
If \code{NULL}, features with missing values are excluded.}

\item{pca_threshold}{the percentile value of the minimum amount of the
variabilities that the selected principal components need to explain}

\item{variance_threshold}{the percentile value of weight each of the 
covariates needs to explain (the rest will be lumped together)}
}
\value{
data frame with weights and factors, combined in a way ready for plotting
}
\description{
prepare the weights of Principal Variance Components
}
\examples{
matrix_test <- example_proteome_matrix[1:150, ]
pvca_df_res <- prepare_PVCA_df(matrix_test, example_sample_annotation, 
technical_factors = c('MS_batch', 'digestion_batch'),
biological_factors = c("Diet", "Sex", "Strain"), 
pca_threshold = .6, variance_threshold = .01, fill_the_missing = -1)
}
