% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proBatch.R
\docType{package}
\name{proBatch}
\alias{proBatch}
\alias{proBatch-package}
\title{proBatch: A package for diagnostics and correction of batch effects, primarily in proteomics}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single sample,
thus it has minimally, `sample_id_col`, `feature_id_column` and `measure_column`,
but usually also `m_score` (in OpenSWATH output result file)}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix,
with feature IDs in rownames and file/sample names as colnames.
Usually the log transformed version of the original data}

\item{sample_annotation}{data matrix with 1) `sample_id_col` (this can be repeated as row names) 2) biological and 3) technical covariates (batches etc)}

\item{sample_id_col}{name of the column in sample_annotation file,
where the filenames (colnames of the data matrix are found)}

\item{batch_column}{column in `sample_annotation` that should be used for batch comparison}

\item{measure_column}{if `df_long` is among the parameters, it is the column
with expression/abundance/intensity; otherwise, it is used internally for consistency}
}
\description{
It adresses the following needs:
\itemize{
\item prepare the original data (e.g. OpenSWATH output matrix and sample annotation file) for analysis.
However, you might need to use `SWATH2stats` additionally
\item Diagnose batch effects, sample-wide and feature-level
\item Correct for batch effects (normalize the data). Other useful package for this purpose is `Normalyzer`.
}
}
\details{
#' To learn more about dplyr, start with the vignettes:
`browseVignettes(package = "proBatch")`
}
\section{common arguments to the functions}{

}

