% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures.R
\name{ProBatchFeatures}
\alias{ProBatchFeatures}
\title{Construct a ProBatchFeatures object from a wide matrix + sample annotation}
\usage{
ProBatchFeatures(
  data_matrix,
  sample_annotation = NULL,
  sample_id_col = "FullRunName",
  name = "raw"
)
}
\arguments{
\item{data_matrix}{numeric matrix (features x samples) or data.frame}

\item{sample_annotation}{data.frame with sample metadata (rows = samples)}

\item{sample_id_col}{character(1), column in sample_annotation that must match colnames(data_matrix).
If missing, rownames(sample_annotation) are used.}

\item{name}{character(1) name of the initial assay added to the object (default: "raw").}
}
\value{
A `ProBatchFeatures` object.
}
\description{
Construct a ProBatchFeatures object from a wide matrix + sample annotation
}
