% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProBatchFeatures.R
\name{ProBatchFeatures_from_long}
\alias{ProBatchFeatures_from_long}
\title{Construct from LONG df via proBatch::long_to_matrix}
\usage{
ProBatchFeatures_from_long(
  df_long,
  sample_annotation = NULL,
  feature_id_col = "peptide_group_label",
  sample_id_col = "FullRunName",
  measure_col = "Intensity",
  level = "feature",
  name = NULL
)
}
\description{
Construct from LONG df via proBatch::long_to_matrix
}
