% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{remove_peptides_with_missing_batch}
\alias{remove_peptides_with_missing_batch}
\title{remove peptides that are missing in the whole batch
useful for some downstream functions as ComBat normalization, that would "choke"}
\usage{
remove_peptides_with_missing_batch(proteome, batch_column = "MS_batch.final",
  feature_id_column = "peptide_group_label")
}
\value{
data frame free of peptides that were not detected across all batches
}
\description{
remove peptides that are missing in the whole batch
useful for some downstream functions as ComBat normalization, that would "choke"
}
