% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{sample_annotation_to_colors}
\alias{sample_annotation_to_colors}
\title{Generate colors for sample annotation}
\usage{
sample_annotation_to_colors(sample_annotation,
  columns_for_plotting = NULL, sample_id_col = "FullRunName",
  factor_columns = c("MS_batch", "EarTag", "Strain", "Diet", "Sex"),
  not_factor_columns = "DateTime", numeric_columns = "order",
  rare_categories_to_other = TRUE, numeric_palette_type = "brewer",
  granularity = 10)
}
\arguments{
\item{sample_annotation}{data frame with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }. 
See "example_sample_annotation" for example with the detailed description (to call it, use \code{help("example_sample_annotation")})}

\item{columns_for_plotting}{only consider these 
columns from sample_annotation}

\item{sample_id_col}{name of the column in \code{sample_annotation} table, where the
filenames (colnames of the data matrix are found).}

\item{factor_columns}{columns of sample_annotation to be 
treated as factors. Note that 
factor and character columns are treated as factors by default.}

\item{not_factor_columns}{don't treat these columns as factors. 
This can be used to override the default behaviour of 
considering factors and 
character columns as factors.}

\item{numeric_columns}{columns of sample_annotation to be 
treated as continuous numeric values.}

\item{rare_categories_to_other}{if True rare categories 
will be merged as 'other'}

\item{numeric_palette_type}{palette to be used for 
numeric values coloring}

\item{granularity}{number of colors to map to the number 
vector (equally spaced between minimum and maximum)}
}
\value{
list of three items: 1) list of colors; 2) data frame of colors; 3) updated sample annotation (e.g. with rare factor levels merged into "other")
}
\description{
Convert the sample annotation data frame to list of colors
the list is named as columns included to use in plotting functions
}
\examples{
color_scheme <- sample_annotation_to_colors (example_sample_annotation, 
factor_columns = c('MS_batch','EarTag', "Strain", 
"Diet", "digestion_batch", "Sex"),
not_factor_columns = 'DateTime',
numeric_columns = c('order'))
}
