% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{sample_annotation_to_colors}
\alias{sample_annotation_to_colors}
\title{convert the sample annotation data frame to list of colors
the list is named as columns included to use in potting functions}
\usage{
sample_annotation_to_colors(sample_annotation, columns_for_plotting = NULL,
  sample_id_column = NULL, factor_columns = NULL,
  not_factor_columns = NULL, rare_categories_to_other = T,
  numerics_to_log = F, numeric_palette_type = "brewer", granularity = 10)
}
\arguments{
\item{sample_annotation}{}

\item{granularity}{number of colors to map to the number vector (equally spaced between minimum and maximum)}
}
\value{
list of colors
}
\description{
convert the sample annotation data frame to list of colors
the list is named as columns included to use in potting functions
}
