% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{sample_annotation_to_colors}
\alias{sample_annotation_to_colors}
\title{Generate colors for sample annotation}
\usage{
sample_annotation_to_colors(sample_annotation,
  columns_for_plotting = NULL, sample_id_col = "FullRunName",
  factor_columns = c("subtype", "caseControl"),
  not_factor_columns = c("RunDate", "ProteinPrepDate"),
  rare_categories_to_other = T, numeric_palette_type = "brewer",
  granularity = 10)
}
\arguments{
\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{columns_for_plotting}{only consider these columns from sample_annotation}

\item{sample_id_col}{name of the column in sample_annotation file, where the
filenames (colnames of the data matrix are found)}

\item{factor_columns}{columns of sample_annotation to be treated as factors. Note that factor and character columns are treated as factors by default.}

\item{not_factor_columns}{don't treat these columns as factors. This can be used to override the default behaviour of considering factors and character columns as factors.}

\item{rare_categories_to_other}{if True rare categories will be merged as 'other'}

\item{numeric_palette_type}{palette to be used for numeric values coloring}

\item{granularity}{number of colors to map to the number vector (equally spaced between minimum and maximum)}
}
\value{
list of colors
}
\description{
Convert the sample annotation data frame to list of colors
the list is named as columns included to use in potting functions
}
