% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_random_peptides.R
\name{sample_random_peptides}
\alias{sample_random_peptides}
\title{sample random peptides for diagnostics}
\usage{
sample_random_peptides(proteome, seed = 1, pep_per_group = 3,
  groups_RT = 10, groups_intensity = 5)
}
\arguments{
\item{proteome}{required columns: \enumerate{
\item \code{m_score}
\item \code{Intensity}
\item \code{peptide_group_label}
\item \code{RT}}}

\item{pep_per_group}{number of peptides to sample per group}

\item{groups_intensity}{}
}
\description{
sample random peptides for diagnostics
}
