% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{summarize_peptides}
\alias{summarize_peptides}
\title{summarize peptides by sample (ranking) and on the contrary, across peptide-wise across samples}
\usage{
summarize_peptides(proteome, sample_id = "FullRunName",
  feature_id = "peptide_group_label")
}
\arguments{
\item{proteome}{}
}
\description{
summarize peptides by sample (ranking) and on the contrary, across peptide-wise across samples
}
