% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_proteome.R
\name{summarize_peptides}
\alias{summarize_peptides}
\title{Summarize run features}
\usage{
summarize_peptides(df_long, sample_id_col = "FullRunName",
  feature_id_col = "peptide_group_label", RT_col = "RT",
  measure_col = "Intensity", m_score = "m_score")
}
\value{
a data frame summarizing features in a dataset on a per sample basis.
  The following columns are returned: `RT_mean`, `Int_mean`, `numb_requants`,
  `median_m_score`, `mean_m_score`, `median_good_m_score` (median of
  `m_score` excluding requants)
}
\description{
Summarizes various peptide properties on a per sample basis. By default will
summarize RT, Intensity and m_score. If your feature does not have some of
these set them to NULL when calling.
}
\details{
summarize peptides by sample (ranking) and on the contrary, across
  peptide-wise across samples
}
\seealso{
Other dataset cleaning functions: \code{\link{clean_requants}},
  \code{\link{remove_peptides_with_missing_batch}}
}
\concept{dataset cleaning functions}
\keyword{internal}
