% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_raw.R
\name{transform_raw_data}
\alias{transform_raw_data}
\alias{log_transform_df}
\alias{unlog_df}
\alias{log_transform_dm}
\alias{unlog_dm}
\title{Functions to log transform raw data before normalization and batch correction}
\usage{
log_transform_df(df_long, log_base = 2, offset = 1,
  measure_col = "Intensity")

unlog_df(df_long, log_base = 2, offset = 1,
  measure_col = "Intensity")

log_transform_dm(data_matrix, log_base = 2, offset = 1)

unlog_dm(data_matrix, log_base = 2, offset = 1)
}
\arguments{
\item{df_long}{data frame where each row is a single feature in a single
sample. It minimally has a \code{sample_id_col}, a \code{feature_id_col} 
and a \code{measure_col}, but usually also an \code{m_score} (in OpenSWATH 
output result file). See \code{help("example_proteome")} for more details.}

\item{log_base}{base of the logarithm for transformation}

\item{offset}{small positive number to prevent 0 conversion to \code{-Inf}}

\item{measure_col}{if \code{df_long} is among the parameters, it is the
column with expression/abundance/intensity; otherwise, it is used
internally for consistency.}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. 
See "example_proteome_matrix" for more details (to call the description, 
use \code{help("example_proteome_matrix")})}
}
\value{
`log_transform_df()` returns \code{df_long}-size data frame, with 
\code{measure_col} log transformed; with old value in another column 
called "beforeLog_intensity" if "intensity" was the value of 
\code{measure_col};
`log_transform_dm()` returns \code{data_matrix} format matrix
}
\description{
Functions to log transform raw data before normalization and batch correction

Log transformation of the data

"Unlog" transformation of the data to pre-log form (for quantification, forcing log-transform)
}
\examples{
log_transformed_df <- log_transform_df(example_proteome)

log_transformed_matrix <- log_transform_dm(example_proteome_matrix, 
log_base = 10, offset = 1)
}
