% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_for_annotation.R
\name{warn_unmapped_columns}
\alias{warn_unmapped_columns}
\title{Warn about unmapped columns}
\usage{
warn_unmapped_columns(
  sample_annotation,
  columns_for_color_mapping,
  sample_id_col
)
}
\arguments{
\item{sample_annotation}{data frame containing sample annotations.}

\item{columns_for_color_mapping}{character vector of columns to be mapped.}

\item{sample_id_col}{character, the ID column.}
}
\value{
No return value, called for side effects (warning).
}
\description{
Emit a warning if some columns will not be mapped to colors.
}
